Testnet.Components.Configuration

Documentation

createConfigJson

Arguments

:: HasCallStack

=> MonadIO m

=> MonadThrow m

=> TmpAbsolutePath

-> ShelleyBasedEra era
The era used for generating the hard fork configuration toggle

-> m (KeyMap Value)

Returns JSON encoded hashes of the era, as well as the hard fork configuration toggle.

createConfigJsonNoHash

Arguments

:: ShelleyBasedEra era
The era used for generating the hard fork configuration toggle

-> KeyMap Value

createSPOGenesisAndFiles

Arguments

:: MonadIO m

=> HasCallStack

=> MonadThrow m

=> CardanoTestnetOptions
The options to use

-> GenesisOptions

-> TestnetOnChainParams

-> TmpAbsolutePath

-> m FilePath
Shelley genesis directory

numSeededUTxOKeys :: Int

getByronGenesisHash :: MonadIO m => MonadThrow m => FilePath -> m (KeyMap Value)

getShelleyGenesisHash :: MonadIO m => FilePath -> Text -> m (KeyMap Value)

getDefaultAlonzoGenesis :: HasCallStack => MonadThrow m => m AlonzoGenesis

An AlonzoGenesis value that is fit to pass to cardanoTestnet

getDefaultShelleyGenesis

Arguments

:: MonadIO m

=> AnyShelleyBasedEra

-> Word64
The max supply

-> GenesisOptions

-> m ShelleyGenesis

A start time and ShelleyGenesis value that are fit to pass to cardanoTestnet

startTimeOffsetSeconds :: NominalDiffTime

For an unknown reason, CLI commands are a lot slower on Windows than on Linux and MacOS. We need to allow a lot more time to set up a testnet.

anyEraToString :: AnyCardanoEra -> String

anyEraToString (AnyCardanoEra ByronEra) returns "byron"

eraToString :: Pretty (eon era) => eon era -> String

eraToString ByronEra returns "byron"