Testnet.Components.Query

Documentation

data EpochStateView

A read-only mutable pointer to an epoch state, updated automatically

getEpochStateView

Arguments

:: (HasCallStack, MonadResource m, MonadTest m)

=> NodeConfigFile 'In
node Yaml configuration file path

-> SocketPath
node socket path

-> m EpochStateView

Create a background thread listening for new epoch states. New epoch states are available to access through EpochStateView, using query functions. The background thread captures any FoldBlocksError into the shared state, so that consumers (e.g. getEpochStateDetails) can fail immediately with a meaningful error message instead of waiting for the full timeout.

getEpochState :: (HasCallStack, MonadTest m, MonadAssertion m, MonadIO m) => EpochStateView -> m AnyNewEpochState

Get epoch state from the view. If the state isn't available, retry waiting up to 25 seconds. Fails immediately if the background thread encountered an error, or after 25 seconds if not yet initialised.

getSlotNumber

Arguments

:: (HasCallStack, MonadIO m, MonadTest m, MonadAssertion m)

=> EpochStateView

-> m SlotNo
The current slot number

getBlockNumber

Arguments

:: (HasCallStack, MonadIO m, MonadTest m, MonadAssertion m)

=> EpochStateView

-> m BlockNo
The number of last produced block

getEpochStateDetails :: (HasCallStack, MonadAssertion m, MonadTest m, MonadIO m) => EpochStateView -> m (AnyNewEpochState, SlotNo, BlockNo)

Access the current epoch state. Returns immediately if state is already available. Blocks up to 25 seconds waiting for initialisation if the background thread has not yet received any epoch state. Fails immediately if the background thread encountered an error.

getMinDRepDeposit

Arguments

:: (HasCallStack, MonadAssertion m, MonadIO m, MonadTest m)

=> EpochStateView

-> ConwayEraOnwards era

-> m Integer
The minimum deposit

Obtain minimum deposit amount for DRep registration from node

getMinGovActionDeposit

Arguments

:: (HasCallStack, MonadAssertion m, MonadIO m, MonadTest m)

=> EpochStateView

-> ConwayEraOnwards era

-> m Integer
The minimum deposit

Obtain minimum deposit amount for governance action from node

getGovState

Arguments

:: (HasCallStack, MonadAssertion m, MonadIO m, MonadTest m)

=> EpochStateView

-> ConwayEraOnwards era

-> m (ConwayGovState (ShelleyLedgerEra era))
The governance state

Obtain governance state from node (CLI query)

getCurrentEpochNo :: (HasCallStack, MonadAssertion m, MonadIO m, MonadTest m) => EpochStateView -> m EpochNo

Return current-ish epoch number. Because we're using Ledger's NewEpochState, the returned epoch number won't be reflecting the current epoch number during the transiontion between the epochs. In other cases it will be the true number of the current epoch.

getTreasuryValue

Arguments

:: (HasCallStack, MonadAssertion m, MonadIO m, MonadTest m)

=> EpochStateView

-> m Coin
The current value of the treasury

Obtain the current value of the treasury from the node

data TestnetWaitPeriod

Constructors

WaitForEpochs EpochInterval

WaitForBlocks Word64

WaitForSlots Word64

waitForEpochs

Arguments

:: (HasCallStack, MonadTest m, MonadAssertion m, MonadIO m)

=> EpochStateView

-> EpochInterval
Number of epochs to wait

-> m EpochNo
The epoch number reached

Wait for the number of epochs

waitUntilEpoch

Arguments

:: (HasCallStack, MonadIO m, MonadTest m)

=> NodeConfigFile 'In

-> SocketPath

-> EpochNo
Desired epoch

-> m EpochNo
The epoch number reached

Block and wait for the desired epoch.

waitForBlocks

Arguments

:: (HasCallStack, MonadIO m, MonadTest m, MonadAssertion m)

=> EpochStateView

-> Word64
Number of blocks to wait

-> m BlockNo
The block number reached

Wait for the requested number of blocks

retryUntilJustM

Arguments

:: (HasCallStack, MonadIO m, MonadTest m, MonadAssertion m)

=> EpochStateView

-> TestnetWaitPeriod
timeout for an operation

-> m (Maybe a)

-> m a

Retries the action until it returns Just or the timeout is reached

retryUntilM

Arguments

:: (HasCallStack, MonadIO m, MonadTest m, MonadAssertion m, Show a)

=> EpochStateView

-> TestnetWaitPeriod
timeout

-> m a
action to retry

-> (a -> Bool)
predicate that must hold

-> m a

Like retryUntilJustM but takes a plain action and a predicate instead of an action returning Maybe. On timeout, annotates the last value that failed the predicate. Intermediate attempts produce no annotations.

findAllUtxos :: forall era m. (HasCallStack, MonadAssertion m, MonadIO m, MonadTest m) => EpochStateView -> ShelleyBasedEra era -> m (UTxO era)

Retrieve all UTxOs map from the epoch state view.

findUtxosWithAddress

Arguments

:: (HasCallStack, MonadAssertion m, MonadIO m, MonadTest m)

=> EpochStateView

-> ShelleyBasedEra era

-> Text
Address

-> m (UTxO era)

Retrieve utxos from the epoch state view for an address.

findLargestUtxoWithAddress

Arguments

:: (HasCallStack, MonadAssertion m, MonadIO m, MonadTest m)

=> EpochStateView

-> ShelleyBasedEra era

-> Text
Address

-> m (Maybe (TxIn, TxOut CtxUTxO era))

Retrieve a one largest utxo

findLargestMultiAssetUtxoWithAddress

Arguments

:: (HasCallStack, MonadAssertion m, MonadIO m, MonadTest m)

=> EpochStateView

-> ShelleyBasedEra era

-> Text
Address

-> m (Maybe (TxIn, TxOut CtxUTxO era))

Retrieve the largest utxo with a multi-asset

findLargestUtxoForPaymentKey :: (MonadTest m, MonadAssertion m, MonadIO m, HasCallStack) => EpochStateView -> ShelleyBasedEra era -> PaymentKeyInfo -> m TxIn

Retrieve a largest UTxO for a payment key info - a convenience wrapper for findLargestUtxoWithAddress.

checkDRepsNumber

Arguments

:: (HasCallStack, MonadAssertion m, MonadIO m, MonadTest m)

=> EpochStateView

-> ShelleyBasedEra ConwayEra
The era in which the test runs

-> Int

-> m ()

checkDRepsNumber config socket execConfig n wait for the number of DReps being n for two epochs. If this number is not attained before two epochs, the test is failed.

checkDRepState

Arguments

:: (HasCallStack, MonadAssertion m, MonadIO m, MonadTest m)

=> EpochStateView

-> ShelleyBasedEra ConwayEra
The era in which the test runs

-> (Map (Credential 'DRepRole) DRepState -> Maybe a)
A function that checks whether the DRep state is correct or up to date and potentially inspects it.

-> m a

checkDRepState sbe configurationFile socketPath execConfig f This functions helps check properties about the DRep state. It waits up to two epochs for the result of applying f to the DRepState to become Just. If f keeps returning Nothing the test fails. If f returns Just, the contents of the Just are returned.

assertNewEpochState

Arguments

:: forall m era value. (HasCallStack, Show value, Eq value, MonadAssertion m, MonadTest m, MonadIO m)

=> EpochStateView
Current epoch state view. It can be obtained using the getEpochStateView function.

-> ShelleyBasedEra era
The ShelleyBasedEra witness for current era.

-> EpochInterval
The maximum wait time in epochs.

-> Lens' (NewEpochState (ShelleyLedgerEra era)) value
The lens to access the specific value in the epoch state.

-> value
The expected value to check in the epoch state.

-> m ()

Assert that the value pointed by the lens in the epoch state is the same as the expected value or it becomes the same within the maxWait epochs. If the value is not reached within the time frame, the test fails.

getProtocolParams :: (MonadAssertion m, MonadTest m, MonadIO m) => EpochStateView -> ConwayEraOnwards era -> m (LedgerProtocolParameters era)

Return current protocol parameters from the governance state

getGovActionLifetime :: (MonadAssertion m, MonadTest m, MonadIO m) => EpochStateView -> ConwayEraOnwards era -> m EpochInterval

Obtains the govActionLifetime from the protocol parameters. The govActionLifetime or governance action maximum lifetime in epochs is the number of epochs such that a governance action submitted during an epoch e expires if it is still not ratified as of the end of epoch: e + govActionLifetime + 1.

getKeyDeposit :: (MonadAssertion m, MonadTest m, MonadIO m) => EpochStateView -> ConwayEraOnwards era -> m Coin

Obtains the key registration deposit from the protocol parameters.

getAccountsStates :: (MonadAssertion m, MonadTest m, MonadIO m) => EpochStateView -> ShelleyBasedEra era -> m (Map (Credential 'Staking) (AccountState (ShelleyLedgerEra era)))

Returns staking accounts state

getTxIx :: (HasCallStack, MonadTest m) => ShelleyBasedEra era -> TxId -> Coin -> (AnyNewEpochState, SlotNo, BlockNo) -> m (Maybe TxIx)

Returns the transaction index of a transaction with a given amount and ID.