Testnet.Defaults

Description

All Byron and Shelley Genesis related functionality

Documentation

defaultAlonzoGenesis :: Either AlonzoGenesisError AlonzoGenesis

defaultByronProtocolParamsJsonValue :: Value

We need a Byron genesis in order to be able to hardfork to the later Shelley based eras. The values here don't matter as the testnet conditions are ultimately determined by the Shelley genesis.

defaultYamlConfig :: [KeyMap Value]

defaultConwayGenesis :: ConwayGenesis StandardCrypto

defaultShelleyGenesis :: UTCTime -> CardanoTestnetOptions -> ShelleyGenesis StandardCrypto

defaultShelleyGenesisFp :: FilePath

defaultYamlHardforkViaConfig :: AnyCardanoEra -> KeyMap Value

Configuration value that allows you to hardfork to any Cardano era at epoch 0.

defaultMainnetTopology :: NetworkTopology