Testnet.Process.Cli.DRep

Documentation

generateDRepKeyPair

Arguments

:: MonadTest m

=> MonadCatch m

=> MonadIO m

=> HasCallStack

=> ExecConfig
Specifies the CLI execution configuration.

-> FilePath
Base directory path where keys will be stored.

-> String
Name for the subfolder that will be created under work folder to store the output keys.

-> m (KeyPair PaymentKey)

Generates a key pair for a decentralized representative (DRep) using cardano-cli.

Returns the generated PaymentKeyPair containing paths to the verification and signing key files.

generateRegistrationCertificate

Arguments

:: MonadTest m

=> MonadCatch m

=> MonadIO m

=> HasCallStack

=> ExecConfig
Specifies the CLI execution configuration.

-> FilePath
Base directory path where the certificate file will be stored.

-> String
Prefix for the output certificate file name. The extension will be .regcert.

-> KeyPair PaymentKey
Payment key pair associated with the DRep. Can be generated using generateDRepKeyPair.

-> Integer
Deposit amount required for DRep registration. The right amount can be obtained using getMinDRepDeposit.

-> m (File Certificate In)

Generates a registration certificate for a decentralized representative (DRep) using cardano-cli.

Returns the generated File DRepRegistrationCertificate In file path to the registration certificate.

createCertificatePublicationTxBody

Arguments

:: MonadAssertion m

=> MonadTest m

=> MonadCatch m

=> MonadIO m

=> ExecConfig
Specifies the CLI execution configuration.

-> EpochStateView
Current epoch state view for transaction building. It can be obtained using the getEpochStateView function.

-> ShelleyBasedEra era
The Shelley-based era (e.g., ShelleyBasedEraShelley) in which the transaction will be constructed.

-> FilePath
Base directory path where the transaction body file will be stored.

-> String
Prefix for the output transaction body file name. The extension will be .txbody.

-> File Certificate In
The file name of the certificate.

-> PaymentKeyInfo
Payment key information associated with the transaction, as returned by cardanoTestnetDefault.

-> m (File TxBody In)

Composes a certificate publication transaction body (without signing) using cardano-cli.

Returns the generated File TxBody In file path to the transaction body.

generateVoteFiles

Arguments

:: MonadTest m

=> MonadIO m

=> MonadCatch m

=> ExecConfig
Specifies the CLI execution configuration.

-> FilePath
Base directory path where the voting files and directories will be stored.

-> String
Name for the subfolder that will be created under work to store the output voting files.

-> TxId
Transaction ID string of the governance action.

-> Word16
Index of the governance action.

-> [(KeyPair PaymentKey, [Char])]
List of tuples where each tuple contains a PaymentKeyPair representing the DRep key pair and a String representing the vote type (i.e: "yes", "no", or "abstain").

-> m [File VoteFile In]

Generates decentralized representative (DRep) voting files (without signing) using cardano-cli.

Returns a list of generated File VoteFile In representing the paths to the generated voting files. TODO: unify with SPO.generateVoteFiles

createVotingTxBody

Arguments

:: MonadAssertion m

=> MonadTest m

=> MonadCatch m

=> MonadIO m

=> ExecConfig
Specifies the CLI execution configuration.

-> EpochStateView
Current epoch state view for transaction building. It can be obtained using the getEpochStateView function.

-> ShelleyBasedEra era
The Shelley-based era (e.g., ShelleyBasedEraShelley) in which the transaction will be constructed.

-> FilePath
Base directory path where the transaction body file will be stored.

-> String
Prefix for the output transaction body file name. The extension will be .txbody.

-> [File VoteFile In]
List of voting files (File VoteFile In) to include in the transaction, obtained using generateVoteFiles.

-> PaymentKeyInfo
Payment key information associated with the transaction, as returned by cardanoTestnetDefault.

-> m (File TxBody In)

Composes a voting transaction body file using cardano-cli. For the transaction to be valid it needs witnesses corresponding to the spent UTxOs and votes issued (typically these witnesses are cryptographic signatures). This function does not sign the transaction, that can be done with signTx.

Returns the generated File TxBody In file path to the transaction body.

registerDRep

Arguments

:: HasCallStack

=> MonadCatch m

=> MonadIO m

=> MonadTest m

=> MonadAssertion m

=> ExecConfig
Specifies the CLI execution configuration.

-> EpochStateView
Current epoch state view for transaction building. It can be obtained using the getEpochStateView function.

-> ConwayEraOnwards ConwayEra
The conway era onwards witness for the era in which the transaction will be constructed.

-> FilePath
Base directory path where the signed transaction file will be stored.

-> FilePath
Name for the subfolder that will be created under work folder to store the output keys.

-> PaymentKeyInfo
Payment key information associated with the transaction, as returned by cardanoTestnetDefault.

-> m (KeyPair PaymentKey)

Register a Delegate Representative (DRep) using cardano-cli, generating a fresh key pair in the process.

Returns the key pair for the DRep as a PaymentKeyPair.

delegateToDRep

Arguments

:: HasCallStack

=> MonadTest m

=> MonadIO m

=> MonadAssertion m

=> MonadCatch m

=> ExecConfig
Specifies the CLI execution configuration.

-> EpochStateView
Current epoch state view for transaction building. It can be obtained using the getEpochStateView function.

-> ShelleyBasedEra ConwayEra
The Shelley-based era (e.g., ConwayEra) in which the transaction will be constructed.

-> FilePath
Base directory path where generated files will be stored.

-> String
Name for the subfolder that will be created under work folder.

-> PaymentKeyInfo
Wallet that will pay for the transaction.

-> KeyPair StakeKey
Staking key pair used for delegation.

-> KeyPair PaymentKey
Delegate Representative (DRep) key pair (PaymentKeyPair) to which delegate.

-> m ()

Delegate to a Delegate Representative (DRep) by creating and submitting a vote delegation certificate transaction using cardano-cli.

getLastPParamUpdateActionId

Arguments

:: HasCallStack

=> MonadTest m

=> MonadCatch m

=> MonadIO m

=> ExecConfig
Specifies the CLI execution configuration.

-> m (Maybe (String, Word16))

This function obtains the identifier for the last enacted parameter update proposal if any.

If no previous proposal was enacted, the function returns Nothing. If there was a previous enacted proposal, the function returns a tuple with its transaction identifier (as a String) and the action index (as a Word16).

makeActivityChangeProposal

Arguments

:: (HasCallStack, MonadBaseControl IO m, MonadAssertion m, MonadTest m, MonadCatch m, MonadIO m, Typeable era)

=> ExecConfig
Specifies the CLI execution configuration.

-> EpochStateView
Current epoch state view for transaction building. It can be obtained using the getEpochStateView function.

-> ConwayEraOnwards era
The ConwayEraOnwards witness for current era.

-> FilePath
Working directory where the files will be stored

-> Maybe (String, Word16)
The transaction id and the index of the previous governance action if any.

-> EpochInterval
The target DRep activity interval to be set by the proposal.

-> KeyPair StakeKey
registered staking keys

-> PaymentKeyInfo
Wallet that will pay for the transaction.

-> EpochInterval
Number of epochs to wait for the proposal to be registered by the chain.

-> m (TxId, Word16)
The transaction id and the index of the governance action.

Create a proposal to change the DRep activity interval. Return the transaction id and the index of the governance action.