Testnet.Process.Cli.Keys

Documentation

cliAddressKeyGen :: (MonadTest m, MonadCatch m, MonadIO m, HasCallStack) => KeyPair PaymentKey -> m ()

cliNodeKeyGen :: (MonadTest m, MonadCatch m, MonadIO m, HasCallStack) => KeyPair StakePoolKey -> File OperatorCounter Out -> m ()

cliNodeKeyGenVrf :: (MonadTest m, MonadCatch m, MonadIO m, HasCallStack) => KeyPair VrfKey -> m ()

cliNodeKeyGenKes :: (MonadTest m, MonadCatch m, MonadIO m, HasCallStack) => KeyPair KesKey -> m ()

cliStakeAddressKeyGen :: (MonadTest m, MonadCatch m, MonadIO m, HasCallStack) => KeyPair StakeKey -> m ()

execCliStdoutToJson

Arguments

:: (HasCallStack, FromJSON a, MonadTest m, MonadCatch m, MonadIO m)

=> ExecConfig
The configuration with which to call the CLI

-> [String]
The CLI command to execute

-> m a

Call a command of the CLI that returns JSON to stdout. Then parse it, and deserialize it to a Haskell value. Fail the test if a step fails. If your CLI command doesn't support returning JSON to stdout, and needs going through a file instead, probably you should add a similar function to this one.

data OperatorCounter

The OperatorCounter

data ByronDelegationKey

Tag a File that holds a ByronDelegationKey.

data ByronDelegationCert

Tag a File that holds a ByronDelegationKey.

cliKeyGen :: TmpDir -> FilePath -> Integration (File ByronKeyLegacy Out)

cliByronSigningKeyAddress :: TmpDir -> Int -> File ByronKeyLegacy In -> FilePath -> Integration (File ByronAddr Out)