Testnet.Process.Cli.SPO

Documentation

checkStakeKeyRegistered

Arguments

:: (MonadTest m, MonadCatch m, MonadIO m, HasCallStack)

=> TmpAbsolutePath

-> NodeConfigFile 'In

-> SocketPath

-> EpochNo
Termination epoch

-> ExecConfig

-> String
Stake address

-> FilePath
Output file path of stake address info

-> m DelegationsAndRewards

createScriptStakeRegistrationCertificate

Arguments

:: (MonadTest m, MonadCatch m, MonadIO m, HasCallStack)

=> TmpAbsolutePath

-> AnyCardanoEra

-> FilePath
Script file

-> Coin
Registration deposit amount used only in Conway

-> FilePath
Output file path

-> m ()

createScriptStakeDelegationCertificate

Arguments

:: (MonadTest m, MonadCatch m, MonadIO m, HasCallStack)

=> TmpAbsolutePath

-> AnyCardanoEra

-> FilePath
Script file

-> File (VKey StakePoolKey) In
Cold stake pool key

-> FilePath
Output file path

-> m ()

createStakeDelegationCertificate

Arguments

:: (MonadTest m, MonadCatch m, MonadIO m, HasCallStack)

=> TmpAbsolutePath

-> ShelleyBasedEra era

-> File (VKey StakeKey) In
Delegate stake verification key file

-> String
Pool id

-> FilePath

-> m ()

createStakeKeyRegistrationCertificate

Arguments

:: (MonadTest m, MonadCatch m, MonadIO m, HasCallStack)

=> TmpAbsolutePath

-> AnyShelleyBasedEra

-> File (VKey StakeKey) In
Stake verification key file

-> Coin
deposit amount used only in Conway

-> FilePath
Output file path

-> m ()

createStakeKeyDeregistrationCertificate

Arguments

:: (MonadTest m, MonadCatch m, MonadIO m, HasCallStack)

=> TmpAbsolutePath

-> ShelleyBasedEra era

-> File (VKey StakeKey) In
Stake verification key file

-> Coin
deposit amount used only in Conway

-> FilePath
Output file path

-> m ()

registerSingleSpo

Arguments

:: (MonadTest m, MonadCatch m, MonadIO m, HasCallStack)

=> AnyShelleyBasedEra

-> Int
Identifier for stake pool

-> TmpAbsolutePath

-> NodeConfigFile 'In

-> SocketPath

-> EpochNo
Termination epoch

-> Int
Testnet magic

-> Coin
key deposit

-> ExecConfig

-> (TxIn, File (SKey PaymentKey) In, String)

-> m (String, KeyPair StakePoolKey, KeyPair VrfKey)
Result tuple: 1. String: Registered stake pool ID 2. Stake pool cold keys 3. Stake pool VRF keys

Related documentation: https://github.com/input-output-hk/cardano-node-wiki/blob/main/docs/stake-pool-operations/8_register_stakepool.md

generateVoteFiles

Arguments

:: (HasCallStack, MonadTest m, MonadIO m, MonadCatch m)

=> ConwayEraOnwards era
The conway era onwards witness for the era in which the transaction will be constructed.

-> ExecConfig
Specifies the CLI execution configuration.

-> FilePath
Base directory path where the voting files and directories will be stored.

-> String
Name for the subfolder that will be created under work to store the output voting files.

-> TxId
Transaction ID string of the governance action.

-> Word16
Index of the governance action.

-> [(SpoNodeKeys, [Char])]
List of tuples where each tuple contains a SpoNodeKeys representing the SPO keys and a String representing the vote type (i.e: "yes", "no", or "abstain").

-> m [File VoteFile In]

Generates Stake Pool Operator (SPO) voting files, using cardano-cli.

Returns a list of generated File VoteFile In representing the paths to the generated voting files. TODO: unify with DRep.generateVoteFiles