Testnet.Process.Cli.Transaction

Documentation

mkSimpleSpendOutputsOnlyTx

Arguments

:: HasCallStack

=> Typeable era

=> MonadAssertion m

=> MonadTest m

=> MonadCatch m

=> MonadIO m

=> ExecConfig
Specifies the CLI execution configuration.

-> EpochStateView
Current epoch state view for transaction building. It can be obtained using the getEpochStateView function.

-> ShelleyBasedEra era
Witness for the current Cardano era.

-> FilePath
Base directory path where the unsigned transaction file will be stored.

-> String
Prefix for the output unsigned transaction file name. The extension will be .txbody.

-> PaymentKeyInfo
Payment key pair used for paying the transaction.

-> PaymentKeyInfo
Payment key of the recipient of the transaction.

-> Coin
Amount of ADA to transfer (in Lovelace).

-> m (File TxBody In)

Calls cardano-cli to build a simple ADA transfer transaction to transfer to the specified recipient the specified amount of ADA.

Returns the generated File TxBody In file path to the created unsigned transaction file.

mkSpendOutputsOnlyTx

Arguments

:: (HasCallStack, Typeable era, MonadAssertion m, MonadTest m, MonadCatch m, MonadIO m)

=> ExecConfig
Specifies the CLI execution configuration.

-> EpochStateView
Current epoch state view for transaction building. It can be obtained using the getEpochStateView function.

-> ShelleyBasedEra era
Witness for the current Cardano era.

-> FilePath
Base directory path where the unsigned transaction file will be stored.

-> String
Prefix for the output unsigned transaction file name. The extension will be .txbody.

-> PaymentKeyInfo
Payment key pair used for paying the transaction.

-> [(TxOutAddress, Coin, Maybe (File ScriptJSON In))]
List of tuples with transaction output addresses, amounts, and reference scripts.

-> m (File TxBody In)

Calls cardano-cli to build a simple ADA transfer transaction to the specified outputs of the specified amount of ADA. Destination address may be specified as a PaymentKeyInfo or with a script file. For each output, an extra optional script file may be provided, and if provided, the script provided will be published in that output as a reference script.

Returns the generated File TxBody In file path to the created unsigned transaction file.

signTx

Arguments

:: MonadTest m

=> MonadCatch m

=> MonadIO m

=> ExecConfig
Specifies the CLI execution configuration.

-> AnyCardanoEra
Specifies the current Cardano era.

-> FilePath
Base directory path where the signed transaction file will be stored.

-> String
Prefix for the output signed transaction file name. The extension will be .tx.

-> File TxBody In
Transaction body to be signed, obtained using createCertificatePublicationTxBody or similar.

-> [Some KeyPair]
List of key pairs used for signing the transaction.

-> m (File SignedTx In)

Calls cardano-cli to signs a transaction body using the specified key pairs.

This function takes five parameters:

Returns the generated File SignedTx In file path to the signed transaction file.

submitTx

Arguments

:: HasCallStack

=> MonadTest m

=> MonadCatch m

=> MonadIO m

=> ExecConfig
Specifies the CLI execution configuration.

-> AnyCardanoEra
Specifies the current Cardano era.

-> File SignedTx In
Signed transaction to be submitted, obtained using signTx.

-> m ()

Submits a signed transaction using cardano-cli.

failToSubmitTx

Arguments

:: MonadTest m

=> MonadCatch m

=> MonadIO m

=> HasCallStack

=> ExecConfig
Specifies the CLI execution configuration.

-> AnyCardanoEra
Specifies the current Cardano era.

-> File SignedTx In
Signed transaction to be submitted, obtained using signTx.

-> String
Substring of the error to check for to ensure submission failed for the right reason.

-> m ()

Attempts to submit a transaction that is expected to fail using cardano-cli.

If the submission fails (the expected behavior), the function succeeds. If the submission succeeds unexpectedly, it raises a failure message that is meant to be caught by Hedgehog.

retrieveTransactionId

Arguments

:: HasCallStack

=> MonadTest m

=> MonadCatch m

=> MonadIO m

=> ExecConfig
Specifies the CLI execution configuration.

-> File SignedTx In
Signed transaction to be submitted, obtained using signTx.

-> m TxId

Retrieves the transaction ID (governance action ID) from a signed transaction file using cardano-cli.

Returns the transaction ID (governance action ID) as a String.

data SignedTx

data TxBody

data TxOutAddress

Constructors

PubKeyAddress PaymentKeyInfo

ScriptAddress (File ScriptJSON In)
The output will be created at the script address.

data VoteFile