Testnet.Process.RunIO

Documentation

execCli' :: (HasCallStack, MonadIO m) => ExecConfig -> [String] -> m String

execCli_ :: (HasCallStack, MonadIO m) => [String] -> m ()

mkExecConfig

Arguments

:: (HasCallStack, MonadIO m)

=> FilePath

-> Sprocket

-> Int
Network id

-> m ExecConfig

procNode

Arguments

:: HasCallStack

=> [String]
Arguments to the CLI command

-> RIO env CreateProcess
Captured stdout

procKesAgent

Arguments

:: HasCallStack

=> [String]
Arguments to the CLI command

-> RIO env CreateProcess
Captured stdout

Create a CreateProcess describing how to start the kes-agent process and an argument list.

execKesAgentControl_ :: (HasCallStack, MonadIO m) => [String] -> m ()

Run kes-agent-control, discarding return value

procCustom

Arguments

:: HasCallStack

=> FilePath
Path to the binary

-> [String]
Arguments to the CLI command

-> RIO env CreateProcess

Like procFlex, but takes an explicit binary path instead of resolving via package name and environment variable.

procFlex

Arguments

:: HasCallStack

=> String
Cabal package name corresponding to the executable

-> String
Environment variable pointing to the binary to run

-> [String]
Arguments to the CLI command

-> RIO env CreateProcess
Captured stdout

Create a CreateProcess describing how to start a process given the Cabal package name corresponding to the executable, an environment variable pointing to the executable, and an argument list.

The actual executable used will the one specified by the environment variable, but if the environment variable is not defined, it will be found instead by consulting the "plan.json" generated by cabal. It is assumed that the project has already been configured and the executable has been built.

liftIOAnnotated :: (HasCallStack, MonadIO m) => IO a -> m a