Testnet.Property.Run

Documentation

data UserProvidedEnv

Whether the user has provided the path to an existing testnet environment through the --node-env flag, or not. This determines whether the testnet will reuse an existing environment or whether a new one should be created.

Constructors

NoUserProvidedEnv
No user-provided environment, a new environment will be created for the user.

UserProvidedEnv FilePath
The user provided the path to an existing testnet environment through the --node-env flag.

runTestnet :: UserProvidedEnv -> (Conf -> Integration TestnetRuntime) -> IO ()

ignoreOn :: Os -> TestTree -> TestTree

ignoreOnWindows :: String -> Property -> TestTree

ignoreOnMac :: String -> Property -> TestTree

ignoreOnMacAndWindows :: String -> Property -> TestTree

disabled :: String -> Property -> TestTree