Testnet.Runtime

Documentation

data LeadershipSlot

Constructors

LeadershipSlot

Fields

slotNumber :: Int

slotTime :: Text

data NodeLoggingFormat

Constructors

NodeLoggingFormatAsJson

NodeLoggingFormatAsText

data PaymentKeyInfo

Constructors

PaymentKeyInfo

Fields

paymentKeyInfoPair :: PaymentKeyPair

paymentKeyInfoAddr :: Text

data PaymentKeyPair

Constructors

PaymentKeyPair

Fields

paymentVKey :: FilePath

paymentSKey :: FilePath

data StakingKeyPair

Constructors

StakingKeyPair

Fields

stakingVKey :: FilePath

stakingSKey :: FilePath

data TestnetRuntime

Constructors

TestnetRuntime

Fields

configurationFile :: FilePath

shelleyGenesisFile :: FilePath

testnetMagic :: Int

poolNodes :: [PoolNode]

wallets :: [PaymentKeyInfo]

delegators :: [Delegator]

data NodeRuntime

Constructors

NodeRuntime

Fields

nodeName :: String

nodeSprocket :: Sprocket

nodeStdinHandle :: Handle

nodeStdout :: FilePath

nodeStderr :: FilePath

nodeProcessHandle :: ProcessHandle

data PoolNode

Constructors

PoolNode

Fields

poolRuntime :: NodeRuntime

poolKeys :: PoolNodeKeys

data PoolNodeKeys

Constructors

PoolNodeKeys

Fields

poolNodeKeysColdVkey :: FilePath

poolNodeKeysColdSkey :: FilePath

poolNodeKeysVrfVkey :: FilePath

poolNodeKeysVrfSkey :: FilePath

poolNodeKeysStakingVkey :: FilePath

poolNodeKeysStakingSkey :: FilePath

data Delegator

Constructors

Delegator

Fields

paymentKeyPair :: PaymentKeyPair

stakingKeyPair :: StakingKeyPair

allNodes :: TestnetRuntime -> [NodeRuntime]

poolSprockets :: TestnetRuntime -> [Sprocket]

poolNodeStdout :: PoolNode -> FilePath

readNodeLoggingFormat :: String -> Either String NodeLoggingFormat

startNode

Arguments

:: TmpAbsolutePath
The temporary absolute path

-> String
The name of the node

-> Int
Node port

-> Int
Testnet magic

-> [String]
The command --socket-path will be added automatically.

-> ExceptT NodeStartFailure (ResourceT IO) NodeRuntime

Start a node, creating file handles, sockets and temp-dirs.

data ShelleyGenesis c

Constructors

ShelleyGenesis

Fields

sgSystemStart :: !UTCTime

sgNetworkMagic :: !Word32

sgNetworkId :: !Network

sgActiveSlotsCoeff :: !PositiveUnitInterval

sgSecurityParam :: !Word64

sgEpochLength :: !EpochSize

sgSlotsPerKESPeriod :: !Word64

sgMaxKESEvolutions :: !Word64

sgSlotLength :: !NominalDiffTimeMicro

sgUpdateQuorum :: !Word64

sgMaxLovelaceSupply :: !Word64

sgProtocolParams :: !(PParams (ShelleyEra c))

sgGenDelegs :: !(Map (KeyHash 'Genesis c) (GenDelegPair c))

sgInitialFunds :: ListMap (Addr c) Coin

sgStaking :: ShelleyGenesisStaking c

shelleyGenesis :: (MonadTest m, MonadIO m, HasCallStack) => TestnetRuntime -> m (ShelleyGenesis StandardCrypto)

getStartTime :: (MonadTest m, MonadIO m, HasCallStack) => FilePath -> TestnetRuntime -> m UTCTime

fromNominalDiffTimeMicro :: NominalDiffTimeMicro -> NominalDiffTime

startLedgerStateLogging

Arguments

:: forall m. MonadCatch m

=> MonadResource m

=> MonadTest m

=> TestnetRuntime

-> FilePath
tmp workspace directory

-> m ()

Start ledger state logging for the first node in the background. Logs will be placed in workspace directorylogsledger-state.log The logging thread will be cancelled when MonadResource releases all resources.