Testnet.Start.Types

Documentation

data CardanoTestnetOptions

Constructors

CardanoTestnetOptions

Fields

cardanoNodes :: [TestnetNodeOptions]

List of node options. Each option will result in a single node being created.

cardanoNodeEra :: AnyCardanoEra

cardanoEpochLength :: Int

cardanoSlotLength :: Double

cardanoTestnetMagic :: Int

cardanoActiveSlotsCoeff :: Double

cardanoMaxSupply :: Word64

The amount of ADA you are starting your testnet with

cardanoEnableP2P :: Bool

cardanoNodeLoggingFormat :: NodeLoggingFormat

cardanoDefaultTestnetOptions :: CardanoTestnetOptions

data TestnetNodeOptions

Specify a BFT node (Pre-Babbage era only) or an SPO (Shelley era onwards only)

Constructors

SpoTestnetNodeOptions (Maybe NodeConfigurationYaml) [String]
These arguments will be appended to the default set of CLI options when starting the node.

extraSpoNodeCliArgs :: TestnetNodeOptions -> [String]

cardanoDefaultTestnetNodeOptions :: [TestnetNodeOptions]

data NodeLoggingFormat

Constructors

NodeLoggingFormatAsJson

NodeLoggingFormatAsText

newtype Conf

Constructors

Conf

Fields

tempAbsPath :: TmpAbsolutePath

newtype NodeConfigurationYaml

Constructors

NodeConfigurationYaml

Fields

unYamlFilePath :: FilePath

mkConf :: (HasCallStack, MonadTest m) => FilePath -> m Conf

Create a Conf from a temporary absolute path. Logs the argument in the test.