Testnet.TestEnumGenerator

Documentation

genTestType :: Name -> Q [Dec]

Create a datatype with the same constructors as the given type, but with a Test prefix and no arguments. For example, if the input type is Maybe, the output type will be TestMaybe, with constructors TestNothing and TestJust, but TestJust will have no arguments.

genAllConstructorsList :: Name -> Q [Dec]

Generate a declaration with a list of all constructors of a type. For example, if the input type is Maybe, the output will be 'allMaybeConstructors = [Nothing, Just]'. Obviously, this will only work if all constructors have types that can be unified, like is the case with nullary constructors like the ones generated by genNewType.