Cardano.Testnet

Description

This module provides a library interface for initiating a local testnet

Testnets

Start a testnet

cardanoTestnet

Arguments

:: HasCallStack

=> MonadUnliftIO m

=> MonadResource m

=> MonadCatch m

=> MonadFail m

=> CardanoTestnetOptions
The options to use

-> Conf
Path to the test sandbox

-> m TestnetRuntime

Starts a number of nodes, as configured by the value of the cardanoNodes field in the CardanoTestnetOptions argument. Regarding this field, you can either:

Pass a value 'UserProvidedNodeOptions filepath' to specify your own node configuration file. In this case, only 1 node will be started (TODO: allow an arbitrary number of nodes to be started)

Pass value NoUserProvidedData to leave this function to generate the node configuration file. In this, one SPO node will be started, as well as two relay nodes.

No matter the scenario above, this function setups a number of credentials and nodes (SPOs and relays), like this:

├── byron-gen-command
│   └── genesis-keys.00{0,1,2}.key
├── delegate-keys
│   ├── delegate{1,2,3}
│   │   ├── kes.{skey,vkey}
│   │   ├── key.{skey,vkey}
│   │   ├── opcert.{cert,counter}
│   │   └── vrf.{skey,vkey}
│   └── README.md
├── drep-keys
│   ├── drep{1,2,3}
│   │   └── drep.{skey,vkey}
│   └── README.md
├── genesis-keys
│   ├── genesis{1,2,3}
│   │   └── key.{skey,vkey}
│   └── README.md
├── logs
│   ├── node{1,2,3}
│   │   ├── node.pid
|   |   └── {stderr,stdout}.log
│   ├── ledger-epoch-state-diffs.log
│   ├── ledger-epoch-state.log
│   ├── node-20241010121635.log
│   └── node.log -> node-20241010121635.log
├── node-data
│   ├── node{1,2,3}
│   │   ├── db
│   │   │   └── <node database files>
│   │   ├── port
│   │   └── topology.json
├── pools-keys
│   ├── pool1
│   │   ├── byron-delegate.key
│   │   ├── byron-delegation.cert
│   │   ├── cold.{skey,vkey}
│   │   ├── kes.{skey,vkey}
│   │   ├── opcert.{cert,counter}
│   │   ├── staking-reward.{skey,vkey}
│   │   └── vrf.{skey,vkey}
│   └── README.md
├── socket
│   ├── node{1,2,3}
│   │   └── sock
├── stake-delegators
│   ├── delegator{1,2,3}
│   │   ├── payment.{skey,vkey}
│   │   └── staking.{skey,vkey}
├── utxo-keys
│   ├── utxo{1,2,3}
│   │   └── utxo.{addr,skey,vkey}
│   └── README.md
├── {alonzo,byron,conway,shelley}-genesis.json
├── configuration.json
├── current-stake-pools.json
└── module

createAndRunTestnet

Arguments

:: HasCallStack

=> CardanoTestnetOptions
The options to use

-> GenesisOptions

-> Conf
Path to the test sandbox

-> Integration TestnetRuntime

A convenience wrapper around createTestnetEnv and cardanoTestnet

createTestnetEnv :: HasCallStack => MonadIO m => MonadThrow m => CardanoTestnetOptions -> GenesisOptions -> CreateEnvOptions -> Conf -> m ()

retryOnAddressInUseError

Arguments

:: forall m a. HasCallStack

=> MonadIO m

=> ExceptT NodeStartFailure m a
action being retried

-> ExceptT NodeStartFailure m a

Retry an action when NodeAddressAlreadyInUseError gets thrown from an action

Testnet options

data CardanoTestnetOptions

Options which, contrary to GenesisOptions are not implemented by tuning the genesis files.

Constructors

CardanoTestnetOptions

Fields

cardanoNodes :: NonEmpty NodeOption

Options controlling how many nodes to create and of which type.

cardanoNodeEra :: AnyShelleyBasedEra

The era to start at

cardanoMaxSupply :: Word64

The amount of Lovelace you are starting your testnet with (forwarded to shelley genesis) TODO move me to GenesisOptions when https://github.com/IntersectMBO/cardano-cli/pull/874 makes it to cardano-node

cardanoNodeLoggingFormat :: NodeLoggingFormat

cardanoNumDReps :: NumDReps

The number of DReps to generate at creation

cardanoEnableNewEpochStateLogging :: Bool

if epoch state logging is enabled

cardanoEnableTxGenerator :: TxGeneratorSupport

Options regarding support for the tx-generator on the testnet (config generation, execution, etc.)

cardanoOutputDir :: UserProvidedEnv

The output directory where to store files, sockets, and so on. If unset, a temporary directory is used.

data NodeOption

Whether a node should be an SPO or just a relay. The '@String' arguments will be appended to the default options when starting the node.

Constructors

SpoNodeOptions [String]

RelayNodeOptions [String]

cardanoDefaultTestnetNodeOptions :: NonEmpty NodeOption

getDefaultAlonzoGenesis :: HasCallStack => MonadThrow m => m AlonzoGenesis

An AlonzoGenesis value that is fit to pass to cardanoTestnet

getDefaultShelleyGenesis

Arguments

:: MonadIO m

=> AnyShelleyBasedEra

-> Word64
The max supply

-> GenesisOptions

-> m ShelleyGenesis

A start time and ShelleyGenesis value that are fit to pass to cardanoTestnet

Configuration

data Conf

Constructors

Conf

Fields

genesisHashesPolicy :: GenesisHashesPolicy

tempAbsPath :: TmpAbsolutePath

updateTimestamps :: UpdateTimestamps

newtype TmpAbsolutePath

Constructors

TmpAbsolutePath

Fields

unTmpAbsPath :: FilePath

data NodeConfiguration

type NodeConfigurationYaml = File NodeConfiguration InOut

mkConf :: (HasCallStack, MonadTest m) => FilePath -> m Conf

Same as mkConfig except that it renders the path when failing in a property test.

makeLogDir :: TmpAbsolutePath -> FilePath

makeSocketDir :: TmpAbsolutePath -> FilePath

makeTmpBaseAbsPath :: TmpAbsolutePath -> FilePath

EpochState processsing helper functions

maybeExtractGovernanceActionIndex

Arguments

:: HasCallStack

=> TxId
transaction id searched for

-> AnyNewEpochState

-> Maybe Word16

Processes

procChairman

Arguments

:: (MonadTest m, MonadCatch m, MonadIO m, HasCallStack)

=> [String]
Arguments to the CLI command

-> m CreateProcess
Captured stdout

Create a CreateProcess describing how to start the cardano-node-chairman process and an argument list.

Utils

integration :: HasCallStack => Integration () -> Property

waitUntilEpoch

Arguments

:: HasCallStack

=> MonadIO m

=> MonadTest m

=> NodeConfigFile In

-> SocketPath

-> EpochNo
Desired epoch

-> m EpochNo
The epoch number reached

Block and wait for the desired epoch.

waitForEpochs

Arguments

:: HasCallStack

=> MonadTest m

=> MonadAssertion m

=> MonadIO m

=> EpochStateView

-> EpochInterval
Number of epochs to wait

-> m EpochNo
The epoch number reached

Wait for the number of epochs

Runtime

data TestnetRuntime

Constructors

TestnetRuntime

Fields

configurationFile :: !(NodeConfigFile In)

shelleyGenesisFile :: !FilePath

testnetMagic :: !Int

testnetNodes :: ![TestnetNode]

wallets :: ![PaymentKeyInfo]

delegators :: ![Delegator]

testnetSprockets :: TestnetRuntime -> [Sprocket]

spoNodes :: TestnetRuntime -> [TestnetNode]

relayNodes :: TestnetRuntime -> [TestnetNode]

data TestnetNode

Constructors

TestnetNode

Fields

nodeName :: !String

poolKeys :: Maybe SpoNodeKeys

Keys are only present for SPO nodes

nodeIpv4 :: !HostAddress

nodePort :: !PortNumber

nodeSprocket :: !Sprocket

nodeStdinHandle :: !Handle

nodeStdout :: !FilePath

nodeStderr :: !FilePath

nodeProcessHandle :: !ProcessHandle

isTestnetNodeSpo :: TestnetNode -> Bool

nodeSocketPath :: TestnetNode -> SocketPath