Cardano.Testnet

Description

This module provides a library interface for initiating a local testnet

Testnets

Start a testnet

cardanoTestnet

Arguments

:: HasCallStack

=> CardanoTestnetOptions
The options to use. Must be consistent with the genesis files.

-> Conf

-> UTCTime
The starting time. Must be the same as the one in the shelley genesis.

-> ShelleyGenesis StandardCrypto
The shelley genesis to use, for example getDefaultShelleyGenesis from this module. Some fields are overridden by the accompanying CardanoTestnetOptions.

-> AlonzoGenesis
The alonzo genesis to use, for example getDefaultAlonzoGenesis from this module.

-> ConwayGenesis StandardCrypto
The conway genesis to use, for example defaultConwayGenesis.

-> Integration TestnetRuntime

Setup a number of credentials and pools, like this:

├── byron
│   └── genesis.json
├── byron-gen-command
│   └── genesis-keys.00{0,1,2}.key
├── byron.genesis.spec.json
├── configuration.yaml
├── current-stake-pools.json
├── delegate-keys
│   ├── delegate{1,2,3}
│   │   ├── kes.{skey,vkey}
│   │   ├── key.{skey,vkey}
│   │   ├── opcert.{cert,counter}
│   │   └── vrf.{skey,vkey}
│   └── README.md
├── drep-keys
│   └── drep{1,2,3}
│       └── drep.{skey,vkey}
├── genesis.{alonzo,conway}.spec.json
├── genesis-keys
│   ├── genesis{1,2,3}
│   │   └── key.{skey,vkey}
│   └── README.md
├── pools-keys
│   ├── pool{1,2,3}
│   │   ├── byron-delegate.key
│   │   ├── byron-delegation.cert
│   │   ├── cold.{skey,vkey}
│   │   ├── kes.{skey,vkey}
│   │   ├── opcert.{cert,counter}
│   │   ├── staking-reward.{skey,vkey}
│   │   ├── topology.json
│   │   └── vrf.{skey,vkey}
│   └── README.md
├── shelley
│   └── genesis.{alonzo,conway,shelley}.json
├── socket
│   └── pool{1,2,3}
│       └── sock
├── stake-delegators
│   └── delegator{1,2,3}
│       ├── payment.{skey,vkey}
│       └── staking.{skey,vkey}
└─── utxo-keys
    ├── README.md
    └── utxo{1,2,3}
        └── utxo.{addr,skey,vkey}

cardanoTestnetDefault :: HasCallStack => CardanoTestnetOptions -> Conf -> Integration TestnetRuntime

Like cardanoTestnet, but using defaults for all configuration files. See cardanoTestnet for additional documentation.

requestAvailablePortNumbers

Arguments

:: HasCallStack

=> MonadIO m

=> MonadTest m

=> Int
Number of ports to request

-> m [PortNumber]

Request a list of unused port numbers for testnet nodes. This shifts availablePortNumber by maxPortsPerRequest in order to make sure that each node gets an unique port.

Testnet options

data CardanoTestnetOptions

Constructors

CardanoTestnetOptions

Fields

cardanoNodes :: [TestnetNodeOptions]

List of node options. Each option will result in a single node being created.

cardanoNodeEra :: AnyCardanoEra

The era to start at

cardanoEpochLength :: Int

An epoch's duration, in number of slots

cardanoSlotLength :: Double

Slot length, in seconds

cardanoTestnetMagic :: Int

cardanoActiveSlotsCoeff :: Double

cardanoMaxSupply :: Word64

The amount of ADA you are starting your testnet with (forwarded to shelley genesis)

cardanoEnableP2P :: Bool

cardanoNodeLoggingFormat :: NodeLoggingFormat

cardanoNumDReps :: Int

The number of DReps to generate at creation

cardanoEnableNewEpochStateLogging :: Bool

if epoch state logging is enabled

data TestnetNodeOptions

Specify a BFT node (Pre-Babbage era only) or an SPO (Shelley era onwards only)

Constructors

SpoTestnetNodeOptions (Maybe NodeConfigurationYaml) [String]
These arguments will be appended to the default set of CLI options when starting the node.

cardanoDefaultTestnetOptions :: CardanoTestnetOptions

cardanoDefaultTestnetNodeOptions :: [TestnetNodeOptions]

getDefaultAlonzoGenesis :: HasCallStack => MonadTest m => m AlonzoGenesis

An AlonzoGenesis value that is fit to pass to cardanoTestnet

getDefaultShelleyGenesis :: HasCallStack => MonadIO m => MonadTest m => CardanoTestnetOptions -> m (UTCTime, ShelleyGenesis StandardCrypto)

A start time and ShelleyGenesis value that are fit to pass to cardanoTestnet

Configuration

newtype Conf

Constructors

Conf

Fields

tempAbsPath :: TmpAbsolutePath

newtype TmpAbsolutePath

Constructors

TmpAbsolutePath

Fields

unTmpAbsPath :: FilePath

newtype NodeConfigurationYaml

Constructors

NodeConfigurationYaml

Fields

unYamlFilePath :: FilePath

mkConf :: (HasCallStack, MonadTest m) => FilePath -> m Conf

Create a Conf from a temporary absolute path. Logs the argument in the test.

makeLogDir :: TmpAbsolutePath -> FilePath

makeSocketDir :: TmpAbsolutePath -> FilePath

makeTmpBaseAbsPath :: TmpAbsolutePath -> FilePath

EpochState processsing helper functions

maybeExtractGovernanceActionIndex

Arguments

:: HasCallStack

=> TxId
transaction id searched for

-> AnyNewEpochState

-> Maybe Word16

Processes

procChairman

Arguments

:: (MonadTest m, MonadCatch m, MonadIO m, HasCallStack)

=> [String]
Arguments to the CLI command

-> m CreateProcess
Captured stdout

Create a CreateProcess describing how to start the cardano-node-chairman process and an argument list.

Utils

integration :: HasCallStack => Integration () -> Property

waitUntilEpoch

Arguments

:: HasCallStack

=> MonadIO m

=> MonadTest m

=> NodeConfigFile In

-> SocketPath

-> EpochNo
Desired epoch

-> m EpochNo
The epoch number reached

Block and wait for the desired epoch.

waitForEpochs

Arguments

:: HasCallStack

=> MonadTest m

=> MonadAssertion m

=> MonadIO m

=> EpochStateView

-> EpochInterval
Number of epochs to wait

-> m EpochNo
The epoch number reached

Wait for the number of epochs

Runtime

data NodeRuntime

Constructors

NodeRuntime

Fields

nodeName :: !String

nodeIpv4 :: !Text

nodePort :: !PortNumber

nodeSprocket :: !Sprocket

nodeStdinHandle :: !Handle

nodeStdout :: !FilePath

nodeStderr :: !FilePath

nodeProcessHandle :: !ProcessHandle

allNodes :: TestnetRuntime -> [NodeRuntime]