Testnet.Components.Configuration

Documentation

createConfigJson

Arguments

:: (MonadTest m, MonadIO m, HasCallStack)

=> TmpAbsolutePath

-> ShelleyBasedEra era
The era used for generating the hard fork configuration toggle

-> m ByteString

Returns JSON encoded hashes of the era, as well as the hard fork configuration toggle.

createSPOGenesisAndFiles

Arguments

:: (MonadTest m, MonadCatch m, MonadIO m, HasCallStack)

=> CardanoTestnetOptions
The options to use

-> GenesisOptions

-> UserProvidedData ShelleyGenesis

-> UserProvidedData AlonzoGenesis

-> UserProvidedData ConwayGenesis

-> TmpAbsolutePath

-> m FilePath
Shelley genesis directory

mkTopologyConfig :: Int -> [Int] -> Int -> Bool -> ByteString

numSeededUTxOKeys :: Int

getByronGenesisHash :: (MonadTest m, MonadIO m) => FilePath -> m (KeyMap Value)

getShelleyGenesisHash :: (MonadTest m, MonadIO m) => FilePath -> Text -> m (KeyMap Value)

getDefaultAlonzoGenesis :: HasCallStack => MonadTest m => ShelleyBasedEra era -> m AlonzoGenesis

An AlonzoGenesis value that is fit to pass to cardanoTestnet

getDefaultShelleyGenesis

Arguments

:: HasCallStack

=> MonadIO m

=> MonadTest m

=> AnyShelleyBasedEra

-> Word64
The max supply

-> GenesisOptions

-> m ShelleyGenesis

A start time and ShelleyGenesis value that are fit to pass to cardanoTestnet

anyEraToString :: AnyCardanoEra -> String

anyEraToString (AnyCardanoEra ByronEra) returns "byron"

eraToString :: Pretty (eon era) => eon era -> String

eraToString ByronEra returns "byron"