Testnet.Components.Configuration

Documentation

createConfigJson

Arguments

:: (MonadTest m, MonadIO m, HasCallStack)

=> TmpAbsolutePath

-> AnyCardanoEra
The era used for generating the hard fork configuration toggle

-> m ByteString

Returns JSON encoded hashes of the era, as well as the hard fork configuration toggle.

createSPOGenesisAndFiles

Arguments

:: (MonadTest m, MonadCatch m, MonadIO m, HasCallStack)

=> NumPools
The number of pools to make

-> NumDReps
The number of pools to make

-> AnyCardanoEra
The era to use

-> ShelleyGenesis StandardCrypto
The shelley genesis to use.

-> AlonzoGenesis
The alonzo genesis to use, for example getDefaultAlonzoGenesis from this module.

-> ConwayGenesis StandardCrypto
The conway genesis to use, for example defaultConwayGenesis.

-> TmpAbsolutePath

-> m FilePath
Shelley genesis directory

mkTopologyConfig :: Int -> [Int] -> Int -> Bool -> ByteString

numSeededUTxOKeys :: Int

getByronGenesisHash :: (MonadTest m, MonadIO m) => FilePath -> m (KeyMap Value)

getShelleyGenesisHash :: (MonadTest m, MonadIO m) => FilePath -> Text -> m (KeyMap Value)

data NumPools

numPools :: CardanoTestnetOptions -> NumPools

data NumDReps

numDReps :: CardanoTestnetOptions -> NumDReps

anyEraToString :: AnyCardanoEra -> String

anyEraToString (AnyCardanoEra ByronEra) returns "byron"

eraToString :: CardanoEra a -> String

eraToString ByronEra returns "byron"