Testnet.Components.Query

Documentation

data EpochStateView

A read-only mutable pointer to an epoch state, updated automatically

getEpochStateView

Arguments

:: HasCallStack

=> MonadResource m

=> MonadTest m

=> MonadCatch m

=> NodeConfigFile In
node Yaml configuration file path

-> SocketPath
node socket path

-> m EpochStateView

Create a background thread listening for new epoch states. New epoch states are available to access through EpochStateView, using query functions.

getEpochState :: HasCallStack => MonadTest m => MonadAssertion m => MonadIO m => EpochStateView -> m AnyNewEpochState

Get epoch state from the view. If the state isn't available, retry waiting up to 15 seconds. Fails when the state is not available after 15 seconds.

getSlotNumber

Arguments

:: HasCallStack

=> MonadIO m

=> MonadTest m

=> MonadAssertion m

=> EpochStateView

-> m SlotNo
The current slot number

getBlockNumber

Arguments

:: HasCallStack

=> MonadIO m

=> MonadTest m

=> MonadAssertion m

=> EpochStateView

-> m BlockNo
The number of last produced block

watchEpochStateUpdate

Arguments

:: forall m a. (HasCallStack, MonadIO m, MonadTest m, MonadAssertion m)

=> EpochStateView
The info to access the epoch state

-> EpochInterval
The maximum number of epochs to wait

-> ((AnyNewEpochState, SlotNo, BlockNo) -> m (Maybe a))
The guard function (Just if the condition is met, Nothing otherwise)

-> m (Maybe a)

Watch the epoch state view until the guard function returns Just or the timeout epoch is reached. Executes the guard function every 300ms. Waits for at most maxWait epochs. The function will return the result of the guard function if it is met within the number of epochs, otherwise it will return Nothing.

getMinDRepDeposit

Arguments

:: HasCallStack

=> MonadAssertion m

=> MonadIO m

=> MonadTest m

=> EpochStateView

-> ConwayEraOnwards era

-> m Integer
The minimum deposit

Obtain minimum deposit amount for DRep registration from node

getMinGovActionDeposit

Arguments

:: HasCallStack

=> MonadAssertion m

=> MonadIO m

=> MonadTest m

=> EpochStateView

-> ConwayEraOnwards era

-> m Integer
The minimum deposit

Obtain minimum deposit amount for governance action from node

getGovState

Arguments

:: HasCallStack

=> MonadAssertion m

=> MonadIO m

=> MonadTest m

=> EpochStateView

-> ConwayEraOnwards era

-> m (ConwayGovState (ShelleyLedgerEra era))
The governance state

Obtain governance state from node (CLI query)

getCurrentEpochNo :: HasCallStack => MonadAssertion m => MonadIO m => MonadTest m => EpochStateView -> m EpochNo

Return current-ish epoch number. Because we're using Ledger's NewEpochState, the returned epoch number won't be reflecting the current epoch number during the transiontion between the epochs. In other cases it will be the true number of the current epoch.

data TestnetWaitPeriod

Constructors

WaitForEpochs EpochInterval

WaitForBlocks Word64

WaitForSlots Word64

waitForEpochs

Arguments

:: HasCallStack

=> MonadTest m

=> MonadAssertion m

=> MonadIO m

=> EpochStateView

-> EpochInterval
Number of epochs to wait

-> m EpochNo
The epoch number reached

Wait for the number of epochs

waitUntilEpoch

Arguments

:: HasCallStack

=> MonadIO m

=> MonadTest m

=> NodeConfigFile In

-> SocketPath

-> EpochNo
Desired epoch

-> m EpochNo
The epoch number reached

Block and wait for the desired epoch.

waitForBlocks

Arguments

:: HasCallStack

=> MonadIO m

=> MonadTest m

=> MonadAssertion m

=> MonadCatch m

=> EpochStateView

-> Word64
Number of blocks to wait

-> m BlockNo
The block number reached

Wait for the requested number of blocks

retryUntilJustM

Arguments

:: HasCallStack

=> MonadIO m

=> MonadTest m

=> MonadAssertion m

=> EpochStateView

-> TestnetWaitPeriod
timeout for an operation

-> m (Maybe a)

-> m a

Retries the action until it returns Just or the timeout is reached

findAllUtxos :: forall era m. HasCallStack => MonadAssertion m => MonadIO m => MonadTest m => EpochStateView -> ShelleyBasedEra era -> m (Map TxIn (TxOut CtxUTxO era))

Retrieve all UTxOs map from the epoch state view.

findUtxosWithAddress

Arguments

:: HasCallStack

=> MonadAssertion m

=> MonadIO m

=> MonadTest m

=> EpochStateView

-> ShelleyBasedEra era

-> Text
Address

-> m (Map TxIn (TxOut CtxUTxO era))

Retrieve utxos from the epoch state view for an address.

findLargestUtxoWithAddress

Arguments

:: HasCallStack

=> MonadAssertion m

=> MonadIO m

=> MonadTest m

=> EpochStateView

-> ShelleyBasedEra era

-> Text
Address

-> m (Maybe (TxIn, TxOut CtxUTxO era))

Retrieve a one largest utxo

findLargestUtxoForPaymentKey :: MonadTest m => MonadAssertion m => MonadIO m => HasCallStack => EpochStateView -> ShelleyBasedEra era -> PaymentKeyInfo -> m TxIn

Retrieve a largest UTxO for a payment key info - a convenience wrapper for findLargestUtxoWithAddress.

checkDRepsNumber

Arguments

:: HasCallStack

=> MonadAssertion m

=> MonadIO m

=> MonadTest m

=> EpochStateView

-> ShelleyBasedEra ConwayEra
The era in which the test runs

-> Int

-> m ()

checkDRepsNumber config socket execConfig n wait for the number of DReps being n for two epochs. If this number is not attained before two epochs, the test is failed.

checkDRepState

Arguments

:: HasCallStack

=> MonadAssertion m

=> MonadIO m

=> MonadTest m

=> EpochStateView

-> ShelleyBasedEra ConwayEra
The era in which the test runs

-> (Map (Credential 'DRepRole StandardCrypto) (DRepState StandardCrypto) -> Maybe a)
A function that checks whether the DRep state is correct or up to date and potentially inspects it.

-> m a

checkDRepState sbe configurationFile socketPath execConfig f This functions helps check properties about the DRep state. It waits up to two epochs for the result of applying f to the DRepState to become Just. If f keeps returning Nothing the test fails. If f returns Just, the contents of the Just are returned.

assertNewEpochState

Arguments

:: forall m era value. HasCallStack

=> Show value

=> Eq value

=> MonadAssertion m

=> MonadTest m

=> MonadIO m

=> EpochStateView
Current epoch state view. It can be obtained using the getEpochStateView function.

-> ShelleyBasedEra era
The ShelleyBasedEra witness for current era.

-> EpochInterval
The maximum wait time in epochs.

-> Lens' (NewEpochState (ShelleyLedgerEra era)) value
The lens to access the specific value in the epoch state.

-> value
The expected value to check in the epoch state.

-> m ()

Assert that the value pointed by the lens in the epoch state is the same as the expected value or it becomes the same within the maxWait epochs. If the value is not reached within the time frame, the test fails.

getGovActionLifetime :: (ConwayEraPParams (ShelleyLedgerEra era), MonadAssertion m, MonadTest m, MonadIO m) => EpochStateView -> ConwayEraOnwards era -> m EpochInterval

Obtains the govActionLifetime from the protocol parameters. The govActionLifetime or governance action maximum lifetime in epochs is the number of epochs such that a governance action submitted during an epoch e expires if it is still not ratified as of the end of epoch: e + govActionLifetime + 1.