Testnet.Defaults

Description

All Byron and Shelley Genesis related functionality

Documentation

defaultAlonzoGenesis :: Either AlonzoGenesisError AlonzoGenesis

defaultByronProtocolParamsJsonValue :: Value

We need a Byron genesis in order to be able to hardfork to the later Shelley based eras. The values here don't matter as the testnet conditions are ultimately determined by the Shelley genesis.

defaultYamlConfig :: KeyMap Value

defaultConwayGenesis :: ConwayGenesis

defaultCommitteeKeyPair :: Int -> KeyPair PaymentKey

defaultCommitteeVkeyFp

Arguments

:: Int
The Committee's index (starts at 1)

-> FilePath

defaultCommitteeSkeyFp

Arguments

:: Int
The Committee's index (starts at 1)

-> FilePath

defaultDRepSkeyFp

Arguments

:: Int
The DRep's index (starts at 1)

-> FilePath

The relative path to DRep secret keys in directories created by cardano-testnet

defaultDRepKeyPair

Arguments

:: Int
The DRep's index (starts at 1)

-> KeyPair PaymentKey

The relative path to DRep keys in directories created by cardano-testnet

defaultDelegatorStakeKeyPair :: Int -> KeyPair StakeKey

The relative path to stake delegator key pairs in directories created by cardano-testnet

defaultEra :: ShelleyBasedEra ConwayEra

The only era supported by cardano-testnet for the moment. It's important to keep the era parameterization everywhere, for ease of development when new eras roll out.

defaultNodeName :: Int -> String

The name of a node (which doesn't have to be a SPO)

defaultNodeDataDir :: Int -> String

The relative path of the node data dir, where the database is stored

defaultP2PTopology :: [adr] -> NetworkTopology adr

defaultSpoColdKeyPair :: Int -> KeyPair StakePoolKey

The relative path to SPO keys in directories created by cardano-testnet

defaultSpoColdVKeyFp :: Int -> FilePath

The relative path to SPO cold verification key in directories created by cardano-testnet

defaultSpoColdSKeyFp :: Int -> FilePath

The relative path to SPO cold secret key in directories created by cardano-testnet

defaultSpoKeys :: Int -> SpoNodeKeys

The relative path to SPO key pairs in directories created by cardano-testnet

defaultSpoKeysDir :: Int -> String

The relative path where the SPO keys for the node are stored

defaultSpoName :: Int -> String

The name of a SPO, used in file system operations

defaultShelleyGenesis :: AnyShelleyBasedEra -> UTCTime -> Word64 -> GenesisOptions -> ShelleyGenesis

defaultGenesisFilepath :: CardanoEra a -> FilePath

defaultYamlHardforkViaConfig :: ShelleyBasedEra era -> KeyMap Value

Configuration value that allows you to hardfork to any Cardano era at epoch 0.

defaultMainnetTopology :: NetworkTopology RelayAccessPoint

defaultUtxoKeys :: Int -> KeyPair PaymentKey

The relative path to UTXO keys

plutusV2Script :: Text

plutusV3Script :: Text

Default plutus script that always succeeds

plutusV3SupplementalDatumScript :: Text

Created via: cabal run plutus-scripts-bench -- print SupplementalDatum -o supplemental-datum.plutus

plutusV2StakeScript :: Text

simpleScript :: Text -> Text