Testnet.Defaults

Description

All Byron and Shelley Genesis related functionality

Documentation

defaultAlonzoGenesis :: Either AlonzoGenesisError AlonzoGenesis

defaultByronProtocolParamsJsonValue :: Value

We need a Byron genesis in order to be able to hardfork to the later Shelley based eras. The values here don't matter as the testnet conditions are ultimately determined by the Shelley genesis.

defaultYamlConfig :: [KeyMap Value]

defaultConwayGenesis :: ConwayGenesis StandardCrypto

defaultCommitteeKeyPair :: Int -> KeyPair PaymentKey

defaultCommitteeVkeyFp

Arguments

:: Int
The Committee's index (starts at 1)

-> FilePath

defaultCommitteeSkeyFp

Arguments

:: Int
The Committee's index (starts at 1)

-> FilePath

defaultDRepSkeyFp

Arguments

:: Int
The DRep's index (starts at 1)

-> FilePath

The relative path to DRep secret keys in directories created by cardano-testnet

defaultDRepKeyPair

Arguments

:: Int
The DRep's index (starts at 1)

-> KeyPair PaymentKey

The relative path to DRep keys in directories created by cardano-testnet

defaultDelegatorStakeKeyPair :: Int -> KeyPair StakingKey

The relative path to stake delegator key pairs in directories created by cardano-testnet

defaultSpoColdKeyPair :: Int -> KeyPair SpoColdKey

The relative path to SPO keys in directories created by cardano-testnet

defaultSPOColdVKeyFp :: Int -> FilePath

The relative path to SPO cold verification key in directories created by cardano-testnet

defaultSPOColdSKeyFp :: Int -> FilePath

The relative path to SPO cold secret key in directories created by cardano-testnet

defaultSpoKeys :: Int -> PoolNodeKeys

The relative path to SPO key pairs in directories created by cardano-testnet

defaultShelleyGenesis :: UTCTime -> CardanoTestnetOptions -> ShelleyGenesis StandardCrypto

defaultGenesisFilepath :: CardanoEra a -> FilePath

defaultYamlHardforkViaConfig :: AnyCardanoEra -> KeyMap Value

Configuration value that allows you to hardfork to any Cardano era at epoch 0.

defaultMainnetTopology :: NetworkTopology

plutusV3Script :: Text

Default plutus script that always succeeds