Testnet.Ping

Documentation

pingNode

Arguments

:: MonadIO m

=> TestnetMagic
testnet magic

-> Sprocket
node sprocket

-> m (Either PingClientError ())
() means success

Ping the node once

checkSprocket :: MonadIO m => Sprocket -> m (Either IOException ())

Check if the sprocket can be connected to. Returns an exception thrown during the connection attempt.

waitForSprocket

Arguments

:: MonadIO m

=> DiffTime
timeout

-> DiffTime
interval

-> Sprocket

-> m (Either IOException ())

Wait for sprocket to become ready. Periodically tries to connect to sprocket, with the provided interval. If there was no success within timeout period, return the last exception thrown during a connection attempt.

waitForPortClosed

Arguments

:: MonadIO m

=> DiffTime
timeout

-> DiffTime
check interval

-> PortNumber

-> m Bool
True if port is closed, False if timeout was reached before that

Wait until port gets closed.

type TestnetMagic = Word32

data PingClientError

Constructors

PceDecodingError

Fields

!String

peer string

!DeserialiseFailure

deserialization exception

PceProtocolError

Fields

!String

peer string

!HandshakeFailure

handshake exception

PceVersionNegotiationError

Fields

!String

peer string

![NodeVersion]

requested versions

![NodeVersion]

received node versions

PceException !SomeException