Testnet.Process.Cli.Transaction

Documentation

signTx

Arguments

:: MonadTest m

=> MonadCatch m

=> MonadIO m

=> ExecConfig
Specifies the CLI execution configuration.

-> AnyCardanoEra
Specifies the current Cardano era.

-> FilePath
Base directory path where the signed transaction file will be stored.

-> String
Prefix for the output signed transaction file name. The extension will be .tx.

-> File TxBody In
Transaction body to be signed, obtained using createCertificatePublicationTxBody or similar.

-> [SomeKeyPair]
List of key pairs used for signing the transaction.

-> m (File SignedTx In)

Calls cardano-cli to signs a transaction body using the specified key pairs.

This function takes five parameters:

Returns the generated File SignedTx In file path to the signed transaction file.

submitTx

Arguments

:: HasCallStack

=> MonadTest m

=> MonadCatch m

=> MonadIO m

=> ExecConfig
Specifies the CLI execution configuration.

-> AnyCardanoEra
Specifies the current Cardano era.

-> File SignedTx In
Signed transaction to be submitted, obtained using signTx.

-> m ()

Submits a signed transaction using cardano-cli.

failToSubmitTx

Arguments

:: MonadTest m

=> MonadCatch m

=> MonadIO m

=> HasCallStack

=> ExecConfig
Specifies the CLI execution configuration.

-> AnyCardanoEra
Specifies the current Cardano era.

-> File SignedTx In
Signed transaction to be submitted, obtained using signTx.

-> String
Substring of the error to check for to ensure submission failed for the right reason.

-> m ()

Attempts to submit a transaction that is expected to fail using cardano-cli.

If the submission fails (the expected behavior), the function succeeds. If the submission succeeds unexpectedly, it raises a failure message that is meant to be caught by Hedgehog.

retrieveTransactionId

Arguments

:: HasCallStack

=> MonadTest m

=> MonadCatch m

=> MonadIO m

=> ExecConfig
Specifies the CLI execution configuration.

-> File SignedTx In
Signed transaction to be submitted, obtained using signTx.

-> m String

Retrieves the transaction ID (governance action ID) from a signed transaction file using cardano-cli.

Returns the transaction ID (governance action ID) as a String.

data SignedTx

data TxBody

data VoteFile