Testnet.Runtime

Documentation

startNode

Arguments

:: HasCallStack

=> MonadResource m

=> MonadCatch m

=> MonadFail m

=> TmpAbsolutePath
The temporary absolute path

-> String
The name of the node

-> HostAddress
Node IPv4 address

-> PortNumber
Node port

-> Int
Testnet magic

-> [String]
The command to execute to start the node. --socket-path, --port, and --host-addr gets added automatically.

-> ExceptT NodeStartFailure m TestnetNode

Start a node, creating file handles, sockets and temp-dirs.

If the port in the function argument was obtained using randomPort which binds to the port first and then closes it, on some operating systems, like MacOS, the port can get stuck in TIME_WAIT state for a significant period. Unfortunately there is no Haskell API giving the ability to check that - this means that the user of this function needs to retry on NodeAddressAlreadyInUseError until this function succeeds. (see state diagram in https://www.rfc-editor.org/rfc/rfc793#section-3.2 p. 23.)

startLedgerNewEpochStateLogging

Arguments

:: HasCallStack

=> MonadResource m

=> TestnetRuntime

-> FilePath
tmp workspace directory

-> m ()

Start ledger's new epoch state logging for the first node in the background. Pretty JSON logs will be placed in: 1. workspace directorylogsledger-new-epoch-state.log 2. workspace directorylogsledger-new-epoch-state-diffs.log NB: The diffs represent the the changes in the NewEpochState between each block or turn of the epoch. We have excluded the stashedAVVMAddresses field of NewEpochState in the JSON rendering. The logging thread will be cancelled when MonadResource releases all resources. Idempotent.

data NodeStartFailure

Constructors

ProcessRelatedFailure ProcessError

ExecutableRelatedFailure SomeException

FileRelatedFailure IOException

NodeExecutableError (Doc Ann)

NodeAddressAlreadyInUseError (Doc Ann)

MaxSprocketLengthExceededError
NodePortNotOpenError IOException

asyncRegister_

Arguments

:: HasCallStack

=> MonadResource m

=> IO a
Action to run in background

-> m (ReleaseKey, Async a)

Runs an action in background, and registers its cancellation to MonadResource.

Orphan instances

(EraTxOut ledgerera, EraGov ledgerera, EraCertState ledgerera, EraStake ledgerera) => ToJSON (NewEpochState ledgerera)

Instance details

Methods

toJSON :: NewEpochState ledgerera -> Value

toEncoding :: NewEpochState ledgerera -> Encoding

toJSONList :: [NewEpochState ledgerera] -> Value

toEncodingList :: [NewEpochState ledgerera] -> Encoding

omitField :: NewEpochState ledgerera -> Bool