Testnet.Runtime

Documentation

startNode

Arguments

:: HasCallStack

=> MonadResource m

=> MonadCatch m

=> MonadFail m

=> TmpAbsolutePath
The temporary absolute path

-> String
The name of the node

-> Text
Node IPv4 address

-> PortNumber
Node port

-> Int
Testnet magic

-> [String]
The command --socket-path will be added automatically.

-> ExceptT NodeStartFailure m NodeRuntime

Start a node, creating file handles, sockets and temp-dirs.

startLedgerNewEpochStateLogging

Arguments

:: HasCallStack

=> MonadCatch m

=> MonadResource m

=> MonadTest m

=> TestnetRuntime

-> FilePath
tmp workspace directory

-> m ()

Start ledger's new epoch state logging for the first node in the background. Pretty JSON logs will be placed in: 1. workspace directorylogsledger-new-epoch-state.log 2. workspace directorylogsledger-new-epoch-state-diffs.log NB: The diffs represent the the changes in the NewEpochState between each block or turn of the epoch. We have excluded the stashedAVVMAddresses field of NewEpochState in the JSON rendering. The logging thread will be cancelled when MonadResource releases all resources. Idempotent.

Orphan instances

(EraTxOut ledgerera, EraGov ledgerera) => ToJSON (NewEpochState ledgerera)

Instance details

Methods

toJSON :: NewEpochState ledgerera -> Value

toEncoding :: NewEpochState ledgerera -> Encoding

toJSONList :: [NewEpochState ledgerera] -> Value

toEncodingList :: [NewEpochState ledgerera] -> Encoding

omitField :: NewEpochState ledgerera -> Bool