Testnet.Start.Cardano

Documentation

data CardanoTestnetCliOptions

Command line options for the cardano-testnet cardano command. Either create a new testnet environment or use a pre-existing one (created by create-env).

Constructors

NoUserProvidedEnv NoUserProvidedEnvOptions

StartFromEnv StartFromEnvOptions

data NoUserProvidedEnvOptions

Options for cardano-testnet cardano when no user-provided environment is given: create a new environment and then start the testnet.

Constructors

NoUserProvidedEnvOptions

Fields

noEnvCreationOptions :: TestnetCreationOptions

noEnvOutputDir :: Maybe FilePath

--output-dir, uses a temporary directory if absent

noEnvRuntimeOptions :: TestnetRuntimeOptions

data StartFromEnvOptions

Options for cardano-testnet cardano --node-env when starting a testnet from a pre-existing environment (created by create-env).

Constructors

StartFromEnvOptions

Fields

fromEnvOptions :: TestnetEnvOptions

fromEnvRuntimeOptions :: TestnetRuntimeOptions

data TestnetCreationOptions

Options for creating a testnet environment (genesis files, topology, ports). Used by both the cardano and create-env subcommands, and by createAndRunTestnet in tests.

Constructors

TestnetCreationOptions

Fields

creationNodes :: TestnetNodesWithOptions

Options controlling how many nodes to create and of which type.

creationEra :: AnyShelleyBasedEra

The era to start at

creationMaxSupply :: Word64

The amount of Lovelace you are starting your testnet with (forwarded to shelley genesis) TODO move me to GenesisOptions when https://github.com/IntersectMBO/cardano-cli/pull/874 makes it to cardano-node

creationNumDReps :: NumDReps

The number of DReps to generate at creation

creationGenesisOptions :: GenesisOptions

creationOnChainParams :: TestnetOnChainParams

data TestnetRuntimeOptions

Options for running testnet nodes (after the environment is created). These are independent of how the environment was created (from scratch or from an existing --node-env path).

Constructors

TestnetRuntimeOptions

Fields

runtimeEnableNewEpochStateLogging :: Bool

if epoch state logging is enabled

runtimeEnableRpc :: RpcSupport

Whether to enable gRPC endpoints in all testnet nodes

runtimeKESSource :: PraosCredentialsSource

runtimeEnableChainStallWatchdog :: Bool

Whether to run a background watchdog that kills the testnet with a diagnosis as soon as the chain has provably stopped producing blocks forever (see Testnet.ChainWatchdog). On by default; disable it only for tests that legitimately halt all block production.

data TestnetEnvOptions

Options specific to the --node-env path: the environment directory and whether to update genesis timestamps before starting.

Constructors

TestnetEnvOptions

Fields

envPath :: FilePath

Path to the pre-existing environment

envUpdateTimestamps :: UpdateTimestamps

data TestnetNodesWithOptions

Specifies the nodes to create for the testnet, split by role (SPO and relay). SPO nodes participate in block production. Relay nodes only forward blocks.

Constructors

TestnetNodesWithOptions

Fields

optSpoNodes :: NonEmpty NodeWithOptions

SPO (stake pool operator) nodes. Must have at least one.

optRelayNodes :: [NodeWithOptions]

Relay (non-producing) nodes

data NodeWithOptions

Configuration specific to each node

Constructors

NodeWithOptions

Fields

nodeBin :: Maybe FilePath

Path to the cardano-node binary to use for running this node. This allows testing different node versions side by side in the same testnet (e.g. to verify compatibility across upgrades). Nothing uses the default resolution: the CARDANO_NODE environment variable if set, otherwise the binary path from cabal's plan.json.

nodeExtraCliArgs :: [String]

Extra CLI arguments passed to cardano-node run

cardanoDefaultTestnetNodesWithOptions :: TestnetNodesWithOptions

data TestnetRuntime

Constructors

TestnetRuntime

Fields

configurationFile :: !(NodeConfigFile 'In)

shelleyGenesisFile :: !FilePath

testnetMagic :: !Int

testnetNodes :: !(NonEmpty TestnetNode)

wallets :: ![PaymentKeyInfo]

delegators :: ![Delegator]

cardanoTestnet

Arguments

:: (HasCallStack, MonadUnliftIO m, MonadResource m, MonadCatch m, MonadFail m)

=> TestnetNodesWithOptions
The nodes to start

-> TestnetRuntimeOptions
Runtime options

-> Conf
Path to the test sandbox

-> m TestnetRuntime

Starts a number of nodes, as given by the first argument. You can either:

Pass a value 'UserProvidedNodeOptions filepath' to specify your own node configuration file. In this case, only 1 node will be started (TODO: allow an arbitrary number of nodes to be started)

Pass value NoUserProvidedData to leave this function to generate the node configuration file. In this, one SPO node will be started, as well as two relay nodes.

No matter the scenario above, this function setups a number of credentials and nodes (SPOs and relays), like this:

├── byron-gen-command
│   └── genesis-keys.00{0,1,2}.key
├── delegate-keys
│   ├── delegate{1,2,3}
│   │   ├── kes.{skey,vkey}
│   │   ├── key.{skey,vkey}
│   │   ├── opcert.{cert,counter}
│   │   └── vrf.{skey,vkey}
│   └── README.md
├── drep-keys
│   ├── drep{1,2,3}
│   │   └── drep.{skey,vkey}
│   └── README.md
├── genesis-keys
│   ├── genesis{1,2,3}
│   │   └── key.{skey,vkey}
│   └── README.md
├── logs
│   ├── node{1,2,3}
│   │   ├── node.pid
|   |   └── {stderr,stdout}.log
│   ├── ledger-epoch-state-diffs.log
│   ├── ledger-epoch-state.log
│   ├── node-20241010121635.log
│   └── node.log -> node-20241010121635.log
├── node-data
│   ├── node{1,2,3}
│   │   ├── db
│   │   │   └── <node database files>
│   │   ├── port
│   │   └── topology.json
├── pools-keys
│   ├── pool1
│   │   ├── byron-delegate.key
│   │   ├── byron-delegation.cert
│   │   ├── cold.{skey,vkey}
│   │   ├── kes.{skey,vkey}
│   │   ├── opcert.{cert,counter}
│   │   ├── staking-reward.{skey,vkey}
│   │   └── vrf.{skey,vkey}
│   └── README.md
├── socket
│   ├── node{1,2,3}
│   │   └── sock
├── stake-delegators
│   ├── delegator{1,2,3}
│   │   ├── payment.{skey,vkey}
│   │   └── staking.{skey,vkey}
├── utxo-keys
│   ├── utxo{1,2,3}
│   │   └── utxo.{addr,skey,vkey}
│   └── README.md
├── {alonzo,byron,conway,shelley}-genesis.json
├── configuration.json
├── current-stake-pools.json
└── module

createAndRunTestnet

Arguments

:: HasCallStack

=> TestnetCreationOptions

-> TestnetRuntimeOptions

-> Conf
Path to the test sandbox

-> Integration TestnetRuntime

A convenience wrapper around createTestnetEnv and cardanoTestnet

createTestnetEnv :: (HasCallStack, MonadIO m, MonadThrow m, MonadFail m) => TestnetCreationOptions -> Conf -> m ()

getDefaultAlonzoGenesis :: (HasCallStack, MonadThrow m) => m AlonzoGenesis

An AlonzoGenesis value that is fit to pass to cardanoTestnet

getDefaultShelleyGenesis

Arguments

:: MonadIO m

=> AnyShelleyBasedEra

-> Word64
The max supply

-> GenesisOptions

-> m ShelleyGenesis

A start time and ShelleyGenesis value that are fit to pass to cardanoTestnet

readNodesWithOptionsFromEnv :: (HasCallStack, MonadIO m) => FilePath -> m TestnetNodesWithOptions

Read node options from an existing testnet environment directory. Scans node-data/ for node directories numbered node1, node2, ... and checks pools-keys/ to classify each as SPO or relay. Validates that nodes are consecutively numbered starting from 1, and that all SPO nodes come before relay nodes.

retryOnAddressInUseError

Arguments

:: forall (m :: Type -> Type) a. (HasCallStack, MonadIO m)

=> ExceptT NodeStartFailure m a
action being retried

-> ExceptT NodeStartFailure m a

Retry an action when NodeAddressAlreadyInUseError gets thrown from an action

liftToIntegration :: HasCallStack => RIO ResourceMap a -> Integration a