Testnet.Start.Cardano

Documentation

data CardanoTestnetCliOptions

Command line options for the cardano-testnet executable. They are used in the parser, and then get split into CardanoTestnetOptions and GenesisOptions

Constructors

CardanoTestnetCliOptions

Fields

cliTestnetOptions :: CardanoTestnetOptions

cliGenesisOptions :: GenesisOptions

data CardanoTestnetOptions

Options which, contrary to GenesisOptions are not implemented by tuning the genesis files.

Constructors

CardanoTestnetOptions

Fields

cardanoNodes :: [TestnetNodeOptions]

List of node options. Each option will result in a single node being created.

cardanoNodeEra :: AnyShelleyBasedEra

The era to start at

cardanoMaxSupply :: Word64

The amount of Lovelace you are starting your testnet with (forwarded to shelley genesis) TODO move me to GenesisOptions when https://github.com/IntersectMBO/cardano-cli/pull/874 makes it to cardano-node

cardanoEnableP2P :: Bool

cardanoNodeLoggingFormat :: NodeLoggingFormat

cardanoNumDReps :: NumDReps

The number of DReps to generate at creation

cardanoEnableNewEpochStateLogging :: Bool

if epoch state logging is enabled

cardanoOutputDir :: Maybe FilePath

The output directory where to store files, sockets, and so on. If unset, a temporary directory is used.

data TestnetNodeOptions

Specify a SPO (Shelley era onwards only) or a Relay node

Constructors

SpoNodeOptions (Maybe NodeConfigurationYaml) [String]

RelayNodeOptions (Maybe NodeConfigurationYaml) [String]
These arguments will be appended to the default set of CLI options when starting the node.

cardanoDefaultTestnetNodeOptions :: [TestnetNodeOptions]

data TestnetRuntime

Constructors

TestnetRuntime

Fields

configurationFile :: !(NodeConfigFile In)

shelleyGenesisFile :: !FilePath

testnetMagic :: !Int

testnetNodes :: ![TestnetNode]

wallets :: ![PaymentKeyInfo]

delegators :: ![Delegator]

cardanoTestnet

Arguments

:: HasCallStack

=> CardanoTestnetOptions
The options to use

-> Conf

-> UserNodeConfig
The node configuration file to use. If omitted it's generated.

-> ShelleyGenesis StandardCrypto
The shelley genesis to use, for example getDefaultShelleyGenesis from this module. Some fields are overridden by the accompanying CardanoTestnetOptions.

-> AlonzoGenesis
The alonzo genesis to use, for example getDefaultAlonzoGenesis from this module.

-> ConwayGenesis StandardCrypto
The conway genesis to use, for example defaultConwayGenesis.

-> Integration TestnetRuntime

Setup a number of credentials and nodes (SPOs and relays), like this:

├── byron-gen-command
│   └── genesis-keys.00{0,1,2}.key
├── delegate-keys
│   ├── delegate{1,2,3}
│   │   ├── kes.{skey,vkey}
│   │   ├── key.{skey,vkey}
│   │   ├── opcert.{cert,counter}
│   │   └── vrf.{skey,vkey}
│   └── README.md
├── drep-keys
│   ├── drep{1,2,3}
│   │   └── drep.{skey,vkey}
│   └── README.md
├── genesis-keys
│   ├── genesis{1,2,3}
│   │   ├── key.{skey,vkey}
│   │   └── key.vkey
│   └── README.md
├── logs
│   ├── node{1,2,3}
│   │   └── {stderr,stdout}.log
│   ├── ledger-epoch-state-diffs.log
│   ├── ledger-epoch-state.log
│   ├── node-20241010121635.log
│   └── node.log -> node-20241010121635.log
├── node-data
│   ├── node{1,2,3}
│   │   ├── db
│   │   │   └── <node database files>
│   │   ├── port
│   │   └── topology.json
├── pools-keys
│   ├── pool1
│   │   ├── byron-delegate.key
│   │   ├── byron-delegation.cert
│   │   ├── cold.{skey,vkey}
│   │   ├── kes.{skey,vkey}
│   │   ├── opcert.{cert,counter}
│   │   ├── staking-reward.{skey,vkey}
│   │   └── vrf.{skey,vkey}
│   └── README.md
├── socket
│   ├── node{1,2,3}
│   │   └── sock
├── stake-delegators
│   ├── delegator{1,2,3}
│   │   ├── payment.{skey,vkey}
│   │   └── staking.{skey,vkey}
├── utxo-keys
│   ├── utxo{1,2,3}
│   │   └── utxo.{addr,skey,vkey}
│   └── README.md
├── alonzo-genesis.json
├── byron.genesis.json
├── byron.genesis.spec.json
├── configuration.yaml
├── conway-genesis.json
├── current-stake-pools.json
├── genesis.{alonzo,conway}.spec.json
├── module
└── shelley-genesis.json

cardanoTestnetDefault :: HasCallStack => CardanoTestnetOptions -> GenesisOptions -> Conf -> Integration TestnetRuntime

Like cardanoTestnet, but using GenesisOptions to obtain the genesis files, instead of passing them directly. See cardanoTestnet for additional documentation.

getDefaultAlonzoGenesis :: HasCallStack => MonadTest m => ShelleyBasedEra era -> m AlonzoGenesis

An AlonzoGenesis value that is fit to pass to cardanoTestnet

getDefaultShelleyGenesis

Arguments

:: HasCallStack

=> MonadIO m

=> MonadTest m

=> AnyShelleyBasedEra

-> Word64
The max supply

-> GenesisOptions

-> m (ShelleyGenesis StandardCrypto)

A start time and ShelleyGenesis value that are fit to pass to cardanoTestnet

retryOnAddressInUseError

Arguments

:: forall m a. HasCallStack

=> MonadTest m

=> MonadIO m

=> ExceptT NodeStartFailure m a
action being retried

-> ExceptT NodeStartFailure m a

Retry an action when NodeAddressAlreadyInUseError gets thrown from an action