Testnet.Start.Cardano

Documentation

data ForkPoint

Constructors

AtVersion Int

AtEpoch Int

data CardanoTestnetOptions

Constructors

CardanoTestnetOptions

Fields

cardanoNodes :: [TestnetNodeOptions]

List of node options. Each option will result in a single node being created.

cardanoNodeEra :: AnyCardanoEra

The era to start at

cardanoEpochLength :: Int

An epoch's duration, in number of slots

cardanoSlotLength :: Double

Slot length, in seconds

cardanoTestnetMagic :: Int

cardanoActiveSlotsCoeff :: Double

cardanoMaxSupply :: Word64

The amount of ADA you are starting your testnet with (forwarded to shelley genesis)

cardanoEnableP2P :: Bool

cardanoNodeLoggingFormat :: NodeLoggingFormat

cardanoNumDReps :: Int

The number of DReps to generate at creation

cardanoEnableNewEpochStateLogging :: Bool

if epoch state logging is enabled

extraSpoNodeCliArgs :: TestnetNodeOptions -> [String]

data TestnetNodeOptions

Specify a BFT node (Pre-Babbage era only) or an SPO (Shelley era onwards only)

Constructors

SpoTestnetNodeOptions (Maybe NodeConfigurationYaml) [String]
These arguments will be appended to the default set of CLI options when starting the node.

cardanoDefaultTestnetOptions :: CardanoTestnetOptions

cardanoDefaultTestnetNodeOptions :: [TestnetNodeOptions]

data TestnetRuntime

Constructors

TestnetRuntime

Fields

configurationFile :: !(NodeConfigFile In)

shelleyGenesisFile :: !FilePath

testnetMagic :: !Int

poolNodes :: ![PoolNode]

wallets :: ![PaymentKeyInfo]

delegators :: ![Delegator]

cardanoTestnet

Arguments

:: HasCallStack

=> CardanoTestnetOptions
The options to use. Must be consistent with the genesis files.

-> Conf

-> UTCTime
The starting time. Must be the same as the one in the shelley genesis.

-> ShelleyGenesis StandardCrypto
The shelley genesis to use, for example getDefaultShelleyGenesis from this module. Some fields are overridden by the accompanying CardanoTestnetOptions.

-> AlonzoGenesis
The alonzo genesis to use, for example getDefaultAlonzoGenesis from this module.

-> ConwayGenesis StandardCrypto
The conway genesis to use, for example defaultConwayGenesis.

-> Integration TestnetRuntime

Setup a number of credentials and pools, like this:

├── byron
│   └── genesis.json
├── byron-gen-command
│   └── genesis-keys.00{0,1,2}.key
├── byron.genesis.spec.json
├── configuration.yaml
├── current-stake-pools.json
├── delegate-keys
│   ├── delegate{1,2,3}
│   │   ├── kes.{skey,vkey}
│   │   ├── key.{skey,vkey}
│   │   ├── opcert.{cert,counter}
│   │   └── vrf.{skey,vkey}
│   └── README.md
├── drep-keys
│   └── drep{1,2,3}
│       └── drep.{skey,vkey}
├── genesis.{alonzo,conway}.spec.json
├── genesis-keys
│   ├── genesis{1,2,3}
│   │   └── key.{skey,vkey}
│   └── README.md
├── pools-keys
│   ├── pool{1,2,3}
│   │   ├── byron-delegate.key
│   │   ├── byron-delegation.cert
│   │   ├── cold.{skey,vkey}
│   │   ├── kes.{skey,vkey}
│   │   ├── opcert.{cert,counter}
│   │   ├── staking-reward.{skey,vkey}
│   │   ├── topology.json
│   │   └── vrf.{skey,vkey}
│   └── README.md
├── shelley
│   └── genesis.{alonzo,conway,shelley}.json
├── socket
│   └── pool{1,2,3}
│       └── sock
├── stake-delegators
│   └── delegator{1,2,3}
│       ├── payment.{skey,vkey}
│       └── staking.{skey,vkey}
└─── utxo-keys
    ├── README.md
    └── utxo{1,2,3}
        └── utxo.{addr,skey,vkey}

cardanoTestnetDefault :: HasCallStack => CardanoTestnetOptions -> Conf -> Integration TestnetRuntime

Like cardanoTestnet, but using defaults for all configuration files. See cardanoTestnet for additional documentation.

getDefaultAlonzoGenesis :: HasCallStack => MonadTest m => m AlonzoGenesis

An AlonzoGenesis value that is fit to pass to cardanoTestnet

getDefaultShelleyGenesis :: HasCallStack => MonadIO m => MonadTest m => CardanoTestnetOptions -> m (UTCTime, ShelleyGenesis StandardCrypto)

A start time and ShelleyGenesis value that are fit to pass to cardanoTestnet

requestAvailablePortNumbers

Arguments

:: HasCallStack

=> MonadIO m

=> MonadTest m

=> Int
Number of ports to request

-> m [PortNumber]

Request a list of unused port numbers for testnet nodes. This shifts availablePortNumber by maxPortsPerRequest in order to make sure that each node gets an unique port.