Testnet.Start.Types

Documentation

data CardanoTestnetCliOptions

Command line options for the cardano-testnet cardano command. Either create a new testnet environment or use a pre-existing one (created by create-env).

Constructors

NoUserProvidedEnv NoUserProvidedEnvOptions

StartFromEnv StartFromEnvOptions

data NoUserProvidedEnvOptions

Options for cardano-testnet cardano when no user-provided environment is given: create a new environment and then start the testnet.

Constructors

NoUserProvidedEnvOptions

Fields

noEnvCreationOptions :: TestnetCreationOptions

noEnvOutputDir :: Maybe FilePath

--output-dir, uses a temporary directory if absent

noEnvRuntimeOptions :: TestnetRuntimeOptions

data StartFromEnvOptions

Options for cardano-testnet cardano --node-env when starting a testnet from a pre-existing environment (created by create-env).

Constructors

StartFromEnvOptions

Fields

fromEnvOptions :: TestnetEnvOptions

fromEnvRuntimeOptions :: TestnetRuntimeOptions

data CardanoTestnetCreateEnvOptions

Command line options for the cardano-testnet create-env subcommand. Creates a sandbox environment without starting nodes.

Constructors

CardanoTestnetCreateEnvOptions

Fields

createEnvCreationOptions :: TestnetCreationOptions

createEnvOutputDir :: FilePath

Required --output directory

data TestnetCreationOptions

Options for creating a testnet environment (genesis files, topology, ports). Used by both the cardano and create-env subcommands, and by createAndRunTestnet in tests.

Constructors

TestnetCreationOptions

Fields

creationNodes :: TestnetNodesWithOptions

Options controlling how many nodes to create and of which type.

creationEra :: AnyShelleyBasedEra

The era to start at

creationMaxSupply :: Word64

The amount of Lovelace you are starting your testnet with (forwarded to shelley genesis) TODO move me to GenesisOptions when https://github.com/IntersectMBO/cardano-cli/pull/874 makes it to cardano-node

creationNumDReps :: NumDReps

The number of DReps to generate at creation

creationGenesisOptions :: GenesisOptions

creationOnChainParams :: TestnetOnChainParams

data TestnetRuntimeOptions

Options for running testnet nodes (after the environment is created). These are independent of how the environment was created (from scratch or from an existing --node-env path).

Constructors

TestnetRuntimeOptions

Fields

runtimeEnableNewEpochStateLogging :: Bool

if epoch state logging is enabled

runtimeEnableRpc :: RpcSupport

Whether to enable gRPC endpoints in all testnet nodes

runtimeKESSource :: PraosCredentialsSource

runtimeEnableChainStallWatchdog :: Bool

Whether to run a background watchdog that kills the testnet with a diagnosis as soon as the chain has provably stopped producing blocks forever (see Testnet.ChainWatchdog). On by default; disable it only for tests that legitimately halt all block production.

data TestnetEnvOptions

Options specific to the --node-env path: the environment directory and whether to update genesis timestamps before starting.

Constructors

TestnetEnvOptions

Fields

envPath :: FilePath

Path to the pre-existing environment

envUpdateTimestamps :: UpdateTimestamps

newtype InputNodeConfigFile

Path to the configuration file of the node, specified by the user

Constructors

InputNodeConfigFile FilePath

newtype NodeId

An abstract node id, used as placeholder in topology files when the actual ports/addresses aren't known yet (i.e. before runtime)

Constructors

NodeId Int

newtype NumDReps

Number of Delegate Represenatives

Constructors

NumDReps Int

newtype NumPools

Number of stake pool nodes

Constructors

NumPools Int

newtype NumRelays

Number of relay nodes

Constructors

NumRelays Int

data RpcSupport

Constructors

RpcDisabled

RpcEnabled

creationNumPools :: TestnetCreationOptions -> NumPools

creationNumRelays :: TestnetCreationOptions -> NumRelays

anyEraToString :: AnyCardanoEra -> String

anyEraToString (AnyCardanoEra ByronEra) returns "byron"

anyShelleyBasedEraToString :: AnyShelleyBasedEra -> String

anyShelleyBasedEraToString $ AnyShelleyBasedEra ShelleyBasedEraConway returns "conway"

defaultTestnetMagic :: Int

The default value for the --testnet-magic option for `cardano-testnet`

eraToString :: Pretty (eon era) => eon era -> String

eraToString ByronEra returns "byron"

data UpdateTimestamps

Constructors

UpdateTimestamps

DontUpdateTimestamps

data TestnetOnChainParams

Constructors

DefaultParams

OnChainParamsFile FilePath
A file path to a JSON file containing on-chain params, formatted as: https://docs.blockfrost.io/#tag/cardano--epochs/GET/epochs/latest

OnChainParamsMainnet

mainnetParamsRequest :: Request

An HTTP request to get a file containing up-to-date mainnet on-chain parameters. The file should be formatted with Blockfrost format: https://docs.blockfrost.io/#tag/cardano--epochs/GET/epochs/latest/parameters

data TestnetNodesWithOptions

Specifies the nodes to create for the testnet, split by role (SPO and relay). SPO nodes participate in block production. Relay nodes only forward blocks.

Constructors

TestnetNodesWithOptions

Fields

optSpoNodes :: NonEmpty NodeWithOptions

SPO (stake pool operator) nodes. Must have at least one.

optRelayNodes :: [NodeWithOptions]

Relay (non-producing) nodes

data NodeWithOptions

Configuration specific to each node

Constructors

NodeWithOptions

Fields

nodeBin :: Maybe FilePath

Path to the cardano-node binary to use for running this node. This allows testing different node versions side by side in the same testnet (e.g. to verify compatibility across upgrades). Nothing uses the default resolution: the CARDANO_NODE environment variable if set, otherwise the binary path from cabal's plan.json.

nodeExtraCliArgs :: [String]

Extra CLI arguments passed to cardano-node run

cardanoDefaultTestnetNodesWithOptions :: TestnetNodesWithOptions

data GenesisOptions

Options that are implemented by writing fields in the Shelley genesis file.

Constructors

GenesisOptions

Fields

genesisTestnetMagic :: Int

genesisEpochLength :: Int

An epoch's duration, in number of slots

genesisSlotLength :: Double

Slot length, in seconds

genesisActiveSlotsCoeff :: Double

data UserProvidedData a

Type used to track whether the user is providing its data (node configuration file path, genesis file, etc.) or whether it needs to be programmatically generated by cardanoTestnet and friends.

Constructors

UserProvidedData a

NoUserProvidedData

data UserProvidedGeneses

Constructors

UserProvidedGeneses

Fields

upgShelleyGenesis :: UserProvidedData ShelleyGenesis

upgAlonzoGenesis :: UserProvidedData AlonzoGenesis

upgConwayGenesis :: UserProvidedData ConwayGenesis

data PraosCredentialsSource

Constructors

UseKesKeyFile

UseKesSocket

data NodeLoggingFormat

Constructors

NodeLoggingFormatAsJson

NodeLoggingFormatAsText

data Conf

Constructors

Conf

Fields

genesisHashesPolicy :: GenesisHashesPolicy

tempAbsPath :: TmpAbsolutePath

updateTimestamps :: UpdateTimestamps

data GenesisHashesPolicy

Constructors

WithHashes

WithoutHashes

data NodeConfiguration

type NodeConfigurationYaml = File NodeConfiguration 'InOut

mkConf :: (HasCallStack, MonadTest m) => FilePath -> m Conf

Same as mkConfig except that it renders the path when failing in a property test.

mkConfigAbs :: FilePath -> IO Conf

Create a Conf from an absolute path, with Genesis Hashes enabled and updating time stamps disabled.