Testnet.Start.Types

Documentation

data CardanoTestnetCliOptions

Command line options for the cardano-testnet executable. They are used in the parser, and then get split into CardanoTestnetOptions and GenesisOptions. If cliNodeEnvironment is provided, don't create a sandbox environment, use the one at the given path.

Constructors

CardanoTestnetCliOptions

Fields

cliTestnetOptions :: CardanoTestnetOptions

cliGenesisOptions :: GenesisOptions

cliNodeEnvironment :: UserProvidedEnv

cliUpdateTimestamps :: UpdateTimestamps

data CardanoTestnetCreateEnvOptions

Constructors

CardanoTestnetCreateEnvOptions

Fields

createEnvTestnetOptions :: CardanoTestnetOptions

createEnvGenesisOptions :: GenesisOptions

createEnvOutputDir :: FilePath

createEnvCreateEnvOptions :: CreateEnvOptions

data CardanoTestnetOptions

Options which, contrary to GenesisOptions are not implemented by tuning the genesis files.

Constructors

CardanoTestnetOptions

Fields

cardanoNodes :: NonEmpty NodeOption

Options controlling how many nodes to create and of which type.

cardanoNodeEra :: AnyShelleyBasedEra

The era to start at

cardanoMaxSupply :: Word64

The amount of Lovelace you are starting your testnet with (forwarded to shelley genesis) TODO move me to GenesisOptions when https://github.com/IntersectMBO/cardano-cli/pull/874 makes it to cardano-node

cardanoNodeLoggingFormat :: NodeLoggingFormat

cardanoNumDReps :: NumDReps

The number of DReps to generate at creation

cardanoEnableNewEpochStateLogging :: Bool

if epoch state logging is enabled

cardanoEnableTxGenerator :: TxGeneratorSupport

Options regarding support for the tx-generator on the testnet (config generation, execution, etc.)

cardanoOutputDir :: UserProvidedEnv

The output directory where to store files, sockets, and so on. If unset, a temporary directory is used.

newtype InputNodeConfigFile

Path to the configuration file of the node, specified by the user

Constructors

InputNodeConfigFile FilePath

newtype NodeId

An abstract node id, used as placeholder in topology files when the actual ports/addresses aren't known yet (i.e. before runtime)

Constructors

NodeId Int

newtype NumDReps

Number of Delegate Represenatives

Constructors

NumDReps Int

newtype NumPools

Number of stake pool nodes

Constructors

NumPools Int

newtype NumRelays

Number of relay nodes

Constructors

NumRelays Int

data TxGeneratorSupport

Constructors

NoTxGeneratorSupport

GenerateTemplateConfigForTxGenerator

GenerateAndRunTxGenerator

cardanoNumPools :: CardanoTestnetOptions -> NumPools

cardanoNumRelays :: CardanoTestnetOptions -> NumRelays

anyEraToString :: AnyCardanoEra -> String

anyEraToString (AnyCardanoEra ByronEra) returns "byron"

anyShelleyBasedEraToString :: AnyShelleyBasedEra -> String

anyShelleyBasedEraToString $ AnyShelleyBasedEra ShelleyBasedEraConway returns "conway"

defaultTestnetMagic :: Int

The default value for the --testnet-magic option for `cardano-testnet`

eraToString :: Pretty (eon era) => eon era -> String

eraToString ByronEra returns "byron"

newtype CreateEnvOptions

Constructors

CreateEnvOptions

Fields

ceoOnChainParams :: TestnetOnChainParams

data UpdateTimestamps

Constructors

UpdateTimestamps

DontUpdateTimestamps

data TestnetOnChainParams

Constructors

DefaultParams

OnChainParamsFile FilePath
A file path to a JSON file containing on-chain params, formatted as: https://docs.blockfrost.io/#tag/cardano--epochs/GET/epochs/latest

OnChainParamsMainnet

mainnetParamsRequest :: Request

An HTTP request to get a file containing up-to-date mainnet on-chain parameters. The file should be formatted with Blockfrost format: https://docs.blockfrost.io/#tag/cardano--epochs/GET/epochs/latest/parameters

data NodeOption

Whether a node should be an SPO or just a relay. The '@String' arguments will be appended to the default options when starting the node.

Constructors

SpoNodeOptions [String]

RelayNodeOptions [String]

isRelayNodeOptions :: NodeOption -> Bool

cardanoDefaultTestnetNodeOptions :: NonEmpty NodeOption

data GenesisOptions

Options that are implemented by writing fields in the Shelley genesis file.

Constructors

GenesisOptions

Fields

genesisTestnetMagic :: Int

genesisEpochLength :: Int

An epoch's duration, in number of slots

genesisSlotLength :: Double

Slot length, in seconds

genesisActiveSlotsCoeff :: Double

data UserProvidedData a

Type used to track whether the user is providing its data (node configuration file path, genesis file, etc.) or whether it needs to be programmatically generated by cardanoTestnet and friends.

Constructors

UserProvidedData a

NoUserProvidedData

data UserProvidedEnv

Constructors

NoUserProvidedEnv

UserProvidedEnv FilePath

data UserProvidedGeneses

Constructors

UserProvidedGeneses

Fields

upgShelleyGenesis :: UserProvidedData ShelleyGenesis

upgAlonzoGenesis :: UserProvidedData AlonzoGenesis

upgConwayGenesis :: UserProvidedData ConwayGenesis

data NodeLoggingFormat

Constructors

NodeLoggingFormatAsJson

NodeLoggingFormatAsText

data Conf

Constructors

Conf

Fields

genesisHashesPolicy :: GenesisHashesPolicy

tempAbsPath :: TmpAbsolutePath

updateTimestamps :: UpdateTimestamps

data GenesisHashesPolicy

Constructors

WithHashes

WithoutHashes

data NodeConfiguration

type NodeConfigurationYaml = File NodeConfiguration InOut

mkConf :: (HasCallStack, MonadTest m) => FilePath -> m Conf

Same as mkConfig except that it renders the path when failing in a property test.

mkConfigAbs :: FilePath -> IO Conf

Create a Conf from an absolute path, with Genesis Hashes enabled and updating time stamps disabled.