Testnet.Start.Types

Documentation

data CardanoTestnetCliOptions

Command line options for the cardano-testnet executable. They are used in the parser, and then get split into CardanoTestnetOptions and GenesisOptions

Constructors

CardanoTestnetCliOptions

Fields

cliTestnetOptions :: CardanoTestnetOptions

cliGenesisOptions :: GenesisOptions

data CardanoTestnetOptions

Options which, contrary to GenesisOptions are not implemented by tuning the genesis files.

Constructors

CardanoTestnetOptions

Fields

cardanoNodes :: TestnetNodeOptions

Options controlling how many nodes to create and whether to use user-provided configuration files, or to generate them automatically.

cardanoNodeEra :: AnyShelleyBasedEra

The era to start at

cardanoMaxSupply :: Word64

The amount of Lovelace you are starting your testnet with (forwarded to shelley genesis) TODO move me to GenesisOptions when https://github.com/IntersectMBO/cardano-cli/pull/874 makes it to cardano-node

cardanoNodeLoggingFormat :: NodeLoggingFormat

cardanoNumDReps :: NumDReps

The number of DReps to generate at creation

cardanoEnableNewEpochStateLogging :: Bool

if epoch state logging is enabled

cardanoOutputDir :: Maybe FilePath

The output directory where to store files, sockets, and so on. If unset, a temporary directory is used.

newtype InputNodeConfigFile

Path to the configuration file of the node, specified by the user

Constructors

InputNodeConfigFile FilePath

newtype NumDReps

Number of Delegate Represenatives

Constructors

NumDReps Int

newtype NumPools

Number of stake pool nodes

Constructors

NumPools Int

newtype NumRelays

Number of relay nodes

Constructors

NumRelays Int

cardanoNumPools :: CardanoTestnetOptions -> NumPools

cardanoNumRelays :: CardanoTestnetOptions -> NumRelays

anyEraToString :: AnyCardanoEra -> String

anyEraToString (AnyCardanoEra ByronEra) returns "byron"

anyShelleyBasedEraToString :: AnyShelleyBasedEra -> String

anyShelleyBasedEraToString $ AnyShelleyBasedEra ShelleyBasedEraConway returns "conway"

eraToString :: Pretty (eon era) => eon era -> String

eraToString ByronEra returns "byron"

data TestnetNodeOptions

Constructors

UserProvidedNodeOptions FilePath
Value used when the user specifies the node configuration file. We start one single SPO node.

AutomaticNodeOptions [AutomaticNodeOption]
Value used when cardano-testnet controls the node configuration files. We start a custom number of nodes.

data AutomaticNodeOption

Type used when the user doesn't specify the node configuration file. We start a custom number of nodes. The '@String' arguments will be appended to the default options when starting the node.

Constructors

SpoNodeOptions [String]

RelayNodeOptions [String]

isRelayNodeOptions :: AutomaticNodeOption -> Bool

cardanoDefaultTestnetNodeOptions :: TestnetNodeOptions

data GenesisOptions

Options that are implemented by writing fields in the Shelley genesis file.

Constructors

GenesisOptions

Fields

genesisTestnetMagic :: Int

genesisEpochLength :: Int

An epoch's duration, in number of slots

genesisSlotLength :: Double

Slot length, in seconds

genesisActiveSlotsCoeff :: Double

data UserProvidedData a

Type used to track whether the user is providing its data (node configuration file path, genesis file, etc.) or whether it needs to be programmatically generated by cardanoTestnet and friends.

Constructors

UserProvidedData a

NoUserProvidedData

data NodeLoggingFormat

Constructors

NodeLoggingFormatAsJson

NodeLoggingFormatAsText

newtype Conf

Constructors

Conf

Fields

tempAbsPath :: TmpAbsolutePath

data NodeConfiguration

type NodeConfigurationYaml = File NodeConfiguration InOut

mkConf :: (HasCallStack, MonadTest m) => FilePath -> m Conf

Create a Conf from a temporary absolute path. Logs the argument in the test.