Testnet.Start.Types

Documentation

data CardanoTestnetOptions

Constructors

CardanoTestnetOptions

Fields

cardanoNodes :: [TestnetNodeOptions]

List of node options. Each option will result in a single node being created.

cardanoNodeEra :: AnyCardanoEra

The era to start at

cardanoEpochLength :: Int

An epoch's duration, in number of slots

cardanoSlotLength :: Double

Slot length, in seconds

cardanoTestnetMagic :: Int

cardanoActiveSlotsCoeff :: Double

cardanoMaxSupply :: Word64

The amount of ADA you are starting your testnet with (forwarded to shelley genesis)

cardanoEnableP2P :: Bool

cardanoNodeLoggingFormat :: NodeLoggingFormat

cardanoNumDReps :: Int

The number of DReps to generate at creation

cardanoEnableNewEpochStateLogging :: Bool

if epoch state logging is enabled

cardanoDefaultTestnetOptions :: CardanoTestnetOptions

anyEraToString :: AnyCardanoEra -> String

anyEraToString (AnyCardanoEra ByronEra) returns "byron"

eraToString :: CardanoEra a -> String

eraToString ByronEra returns "byron"

data TestnetNodeOptions

Specify a BFT node (Pre-Babbage era only) or an SPO (Shelley era onwards only)

Constructors

SpoTestnetNodeOptions (Maybe NodeConfigurationYaml) [String]
These arguments will be appended to the default set of CLI options when starting the node.

extraSpoNodeCliArgs :: TestnetNodeOptions -> [String]

cardanoDefaultTestnetNodeOptions :: [TestnetNodeOptions]

data NodeLoggingFormat

Constructors

NodeLoggingFormatAsJson

NodeLoggingFormatAsText

newtype Conf

Constructors

Conf

Fields

tempAbsPath :: TmpAbsolutePath

newtype NodeConfigurationYaml

Constructors

NodeConfigurationYaml

Fields

unYamlFilePath :: FilePath

mkConf :: (HasCallStack, MonadTest m) => FilePath -> m Conf

Create a Conf from a temporary absolute path. Logs the argument in the test.