Testnet.Types

Documentation

data LeadershipSlot

Constructors

LeadershipSlot

Fields

slotNumber :: Int

slotTime :: Text

data NodeLoggingFormat

Constructors

NodeLoggingFormatAsJson

NodeLoggingFormatAsText

data PaymentKeyInfo

Constructors

PaymentKeyInfo

Fields

paymentKeyInfoPair :: KeyPair PaymentKey

paymentKeyInfoAddr :: Text

data TestnetRuntime

Constructors

TestnetRuntime

Fields

configurationFile :: !(NodeConfigFile In)

shelleyGenesisFile :: !FilePath

testnetMagic :: !Int

testnetNodes :: ![TestnetNode]

wallets :: ![PaymentKeyInfo]

delegators :: ![Delegator]

spoNodes :: TestnetRuntime -> [TestnetNode]

relayNodes :: TestnetRuntime -> [TestnetNode]

testnetSprockets :: TestnetRuntime -> [Sprocket]

data TestnetNode

Constructors

TestnetNode

Fields

nodeName :: !String

poolKeys :: Maybe SpoNodeKeys

Keys are only present for SPO nodes

nodeIpv4 :: !HostAddress

nodePort :: !PortNumber

nodeSprocket :: !Sprocket

nodeStdinHandle :: !Handle

nodeStdout :: !FilePath

nodeStderr :: !FilePath

nodeProcessHandle :: !ProcessHandle

nodeSocketPath :: TestnetNode -> SocketPath

isTestnetNodeSpo :: TestnetNode -> Bool

data SpoNodeKeys

Constructors

SpoNodeKeys

Fields

poolNodeKeysCold :: KeyPair SpoColdKey

poolNodeKeysVrf :: KeyPair VrfKey

poolNodeKeysStaking :: KeyPair StakingKey

data Delegator

Constructors

Delegator

Fields

paymentKeyPair :: KeyPair PaymentKey

stakingKeyPair :: KeyPair StakingKey

data KeyPair k

Constructors

KeyPair

Fields

verificationKey :: forall dir. File (VKey k) dir

signingKey :: forall dir. File (SKey k) dir

verificationKeyFp :: KeyPair k -> FilePath

signingKeyFp :: KeyPair k -> FilePath

data SomeKeyPair

Constructors

forall a. SomeKeyPair (KeyPair a)

data VKey k

Verification key tag

data SKey k

Signing key tag

data ColdPoolKey

data VrfKey

data StakingKey

data PaymentKey

Shelley-era payment keys. Used for Shelley payment addresses and witnessing transactions that spend from these addresses.

This is a type level tag, used with other interfaces like Key.

data DRepKey

data SpoColdKey

readNodeLoggingFormat :: String -> Either String NodeLoggingFormat

data ShelleyGenesis c

Shelley genesis information

Note that this is needed only for a pure Shelley network, hence it being defined here rather than in its own module. In mainnet, Shelley will transition naturally from Byron, and thus will never have its own genesis information.

Constructors

ShelleyGenesis

Fields

sgSystemStart :: !UTCTime

sgNetworkMagic :: !Word32

sgNetworkId :: !Network

sgActiveSlotsCoeff :: !PositiveUnitInterval

sgSecurityParam :: !Word64

sgEpochLength :: !EpochSize

sgSlotsPerKESPeriod :: !Word64

sgMaxKESEvolutions :: !Word64

sgSlotLength :: !NominalDiffTimeMicro

sgUpdateQuorum :: !Word64

sgMaxLovelaceSupply :: !Word64

sgProtocolParams :: !(PParams (ShelleyEra c))

sgGenDelegs :: !(Map (KeyHash 'Genesis c) (GenDelegPair c))

sgInitialFunds :: ListMap (Addr c) Coin

sgInitialFunds is intentionally kept lazy, as it can otherwise cause out-of-memory problems in testing and benchmarking.

sgStaking :: ShelleyGenesisStaking c

sgStaking is intentionally kept lazy, as it can otherwise cause out-of-memory problems in testing and benchmarking.

shelleyGenesis :: MonadTest m => MonadIO m => HasCallStack => TestnetRuntime -> m (ShelleyGenesis StandardCrypto)

getStartTime :: MonadTest m => MonadIO m => HasCallStack => FilePath -> TestnetRuntime -> m UTCTime

testnetDefaultIpv4Address :: HostAddress

Hardcoded testnet IPv4 address pointing to the local host

showIpv4Address :: IsString s => HostAddress -> s

Format IPv4 address as a string-like e.g. 127.0.0.1